perf: modularize qmd_memory.rs into logical components
- Split `src/memory/qmd_memory.rs` into `src/memory/qmd_memory/` module.
- Extract types into `types.rs`.
- Extract constants and utilities into `utils.rs`.
- Extract embedding cache logic into `cache.rs`.
- Extract search and scoring logic into `search.rs`.
- Extract storage mapping into `storage.rs`.
- Extract indexing and document variants into `index.rs`.
- Maintain `QmdMemory` API in `mod.rs` by delegating to new submodules.
- Convert `query_with_embedding` to `async fn` for better integration.
- Ensure all 11 unit tests pass in the new structure.
