--- src/memory/qmd_memory.rs
@@ -1,3219 +1,0 @@
-//! QMD Memory - lightweight in-memory document store with cached search.
-
-use anyhow::Result;
-use chrono::{Datelike, Duration, NaiveDate};
-use regex::Regex;
-use serde::{Deserialize, Serialize};
-use serde_json::{json, Value};
-use sha2::{Digest, Sha256};
-use std::{
-    cmp::Ordering,
-    collections::HashMap,
-    sync::LazyLock,
-    sync::{
-        atomic::{AtomicUsize, Ordering as AtomicOrdering},
-        Arc,
-    },
-    time::Instant,
-};
-use tokio::sync::RwLock as AsyncRwLock;
-
-use crate::memory::schema::{
-    matches_filters, normalize_metadata, resolve_metadata, EvidenceKind, MemoryKind,
-    MemoryQueryFilters, TypedMemoryPayload,
-};
-use crate::memory::surreal_store::{MemoryRecord, MemoryStore};
-use crate::utils::crypto::hex_encode;
-
-/// Compute a stable SHA256 content hash for deduplication.
-#[allow(dead_code)]
-fn _compute_content_hash(content: &str) -> String {
-    hex_encode(Sha256::digest(content.as_bytes()).as_slice())
-}
-static SPEAKER_COLON_RE: LazyLock<Regex> =
-    LazyLock::new(|| Regex::new(r"(?m)^([^:\s]+):\s*").unwrap());
-static SPEAKER_BRACKET_RE: LazyLock<Regex> =
-    LazyLock::new(|| Regex::new(r"(?m)^\[([^]\s]+)\]").unwrap());
-static SPEAKER_ROLE_RE: LazyLock<Regex> = LazyLock::new(|| {
-    Regex::new(
-        r"(?i)(?:Speaker|Person|Host|Guest|Interviewer|Interviewee|Moderator):\s*([A-Z][a-zA-Z]+)",
-    )
-    .unwrap()
-});
-static QUERY_SPEAKER_RE: LazyLock<Regex> = LazyLock::new(|| {
-    Regex::new(r"(?i)\b(?:who|what|where|when|why|how|did|was|were)(?:\s+is|\s+did|\s+was|\s+were)?\s+([A-Z][a-zA-Z]+)").unwrap()
-});
-static SHE_RE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(?i)\bshe\b").unwrap());
-static HE_RE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(?i)\bhe\b").unwrap());
-static SYNONYM_MAP: LazyLock<HashMap<&'static str, &'static [&'static str]>> =
-    LazyLock::new(|| {
-        HashMap::from([
-            ("bug", &["issue", "error", "failure", "defect"][..]),
-            ("cache", &["caching", "memoization", "store"][..]),
-            ("fast", &["quick", "speed", "latency"][..]),
-            ("memory", &["context", "retrieval", "knowledge"][..]),
-            ("search", &["lookup", "find", "retrieve"][..]),
-            ("vector", &["embedding", "semantic", "dense"][..]),
-            ("query", &["question", "request", "prompt"][..]),
-            ("reasoning", &["multi-hop", "inference", "analysis"][..]),
-        ])
-    });
-const RRF_K: f32 = 60.0;
-const KEYWORD_WEIGHT: f32 = 0.7;
-const SEMANTIC_WEIGHT: f32 = 0.3;
-const MAX_EXPANSIONS: usize = 4;
-const MAX_MULTI_HOP_DEPTH: usize = 2;
-const MAX_RERANK_CANDIDATES: usize = 32;
-static DIA_ID_RE: LazyLock<Regex> =
-    LazyLock::new(|| Regex::new(r"(?i)^([a-z]+\d+):0*([0-9]+)$").unwrap());
-static LOCOMO_PATH_DIA_ID_RE: LazyLock<Regex> =
-    LazyLock::new(|| Regex::new(r"(?i)(/)([a-z]+\d+):0*([0-9]+)([#/]|$)").unwrap());
-
-/// CRITICAL FIX: Embedding cache with TTL to avoid re-embedding identical content
-/// Previously, every document was re-embedded on each access (~118ms per document)
-/// Now uses SHA256 hash of preprocessed content as key with 1-hour TTL
-const EMBEDDING_CACHE_TTL_SECS: u64 = 3600; // 1 hour
-
-struct EmbeddingCacheEntry {
-    vector: Vec<f32>,
-    cached_at: Instant,
-}
-
-/// Global embedding cache - shared across all QmdMemory instances
-static EMBEDDING_CACHE: LazyLock<Arc<AsyncRwLock<HashMap<String, EmbeddingCacheEntry>>>> =
-    LazyLock::new(|| Arc::new(AsyncRwLock::new(HashMap::new())));
-
-/// Compute a stable hash key for embedding cache
-fn embedding_cache_key(content: &str) -> String {
-    let mut hasher = Sha256::new();
-    hasher.update(content.as_bytes());
-    hex_encode(&hasher.finalize())
-}
-
-/// Clean up expired entries from the embedding cache
-async fn clean_embedding_cache() {
-    let mut cache = EMBEDDING_CACHE.write().await;
-    let now = Instant::now();
-    cache.retain(|_, entry| {
-        now.duration_since(entry.cached_at).as_secs() < EMBEDDING_CACHE_TTL_SECS
-    });
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct MemoryDocument {
-    pub id: Option<String>,
-    pub path: String,
-    pub content: String,
-    pub metadata: serde_json::Value,
-    #[serde(default)]
-    pub content_vector: Option<Vec<f32>>,
-    pub embedding: Vec<f32>,
-}
-
-impl MemoryDocument {
-    pub fn estimated_bytes(&self) -> u64 {
-        self.id
-            .as_ref()
-            .map(|value| value.len())
-            .unwrap_or_default() as u64
-            + self.path.len() as u64
-            + self.content.len() as u64
-            + self.metadata.to_string().len() as u64
-            + self
-                .content_vector
-                .as_ref()
-                .map(|value| value.len() * std::mem::size_of::<f32>())
-                .unwrap_or_default() as u64
-            + (self.embedding.len() * std::mem::size_of::<f32>()) as u64
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct MemoryUsage {
-    pub document_count: usize,
-    pub storage_bytes: u64,
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct CacheMetrics {
-    pub hits: usize,
-    pub misses: usize,
-    pub entries: usize,
-}
-
-#[derive(Debug, Clone)]
-pub struct CachedSearchResult {
-    pub documents: Vec<MemoryDocument>,
-    pub cache_hit: bool,
-}
-
-/// CRITICAL FIX: Added workspace_id to prevent cross-workspace cache contamination
-/// Previously, searches from different workspaces shared the same cache key,
-/// potentially leaking results from workspace A to workspace B.
-#[derive(Debug, Clone, Hash, PartialEq, Eq)]
-struct SearchCacheKey {
-    workspace_id: String,
-    query: String,
-    limit: usize,
-    filters: String,
-}
-
-#[derive(Default)]
-struct CacheCounters {
-    hits: AtomicUsize,
-    misses: AtomicUsize,
-}
-
-#[derive(Clone)]
-pub struct QmdMemory {
-    workspace_id: String,
-    docs: Arc<AsyncRwLock<Vec<MemoryDocument>>>,
-    search_cache: Arc<AsyncRwLock<HashMap<SearchCacheKey, Vec<MemoryDocument>>>>,
-    cache_counters: Arc<CacheCounters>,
-    store: Arc<AsyncRwLock<Option<Arc<dyn MemoryStore>>>>,
-}
-
-impl QmdMemory {
-    pub fn new(docs: Arc<AsyncRwLock<Vec<MemoryDocument>>>) -> Self {
-        Self::new_with_workspace(docs, "default")
-    }
-
-    pub fn new_with_workspace(
-        docs: Arc<AsyncRwLock<Vec<MemoryDocument>>>,
-        workspace_id: impl Into<String>,
-    ) -> Self {
-        Self {
-            workspace_id: workspace_id.into(),
-            docs,
-            search_cache: Arc::new(AsyncRwLock::new(HashMap::new())),
-            cache_counters: Arc::new(CacheCounters::default()),
-            store: Arc::new(AsyncRwLock::new(None)),
-        }
-    }
-
-    pub fn workspace_id(&self) -> &str {
-        &self.workspace_id
-    }
-
-    pub async fn set_store(&self, store: Arc<dyn MemoryStore>) {
-        *self.store.write().await = Some(store);
-    }
-
-    async fn store(&self) -> Option<Arc<dyn MemoryStore>> {
-        self.store.read().await.clone()
-    }
-
-    /// Load workspace state from persistent store on startup.
-    /// This is CRITICAL for persistence - without this, data written to SurrealDB
-    /// (or file store) before a restart would be lost on restart.
-    pub async fn init(&self) -> Result<()> {
-        if let Some(store) = self.store().await {
-            let state = store.load_workspace_state(&self.workspace_id).await?;
-            let docs: Vec<MemoryDocument> = state
-                .memories
-                .into_iter()
-                .map(|record| record.to_document())
-                .collect();
-            let loaded_memories = docs.len();
-            *self.docs.write().await = docs;
-            tracing::info!(
-                workspace_id = %self.workspace_id,
-                loaded_memories = loaded_memories,
-                "QmdMemory loaded from persistent store"
-            );
-        }
-        Ok(())
-    }
-
-    pub async fn search(&self, query_text: &str, limit: usize) -> Result<Vec<MemoryDocument>> {
-        self.search_filtered(query_text, limit, None).await
-    }
-
-    pub async fn bm25_search(
-        &self,
-        query: &str,
-        limit: usize,
-        filters: Option<&MemoryQueryFilters>,
-    ) -> Result<Vec<MemoryDocument>> {
-        let docs = self.all_documents().await;
-        let filtered_docs: Vec<MemoryDocument> = docs
-            .into_iter()
-            .filter(|doc| matches_filters(&doc.path, &doc.metadata, &self.workspace_id, filters))
-            .collect();
-
-        if filtered_docs.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        let scored = crate::search::bm25::score_documents(
-            query,
-            &filtered_docs,
-            crate::search::bm25::Bm25Params::default(),
-        );
-
-        let mut results = Vec::new();
-        for (_, id) in scored.into_iter().take(limit) {
-            if let Some(doc) = filtered_docs
-                .iter()
-                .find(|d| d.id.as_deref() == Some(&id) || d.path == id)
-            {
-                results.push(doc.clone());
-            }
-        }
-
-        Ok(results)
-    }
-
-    pub async fn search_filtered(
-        &self,
-        query_text: &str,
-        limit: usize,
-        filters: Option<&MemoryQueryFilters>,
-    ) -> Result<Vec<MemoryDocument>> {
-        let optimized = self
-            .search_hybrid_optimized(query_text, limit, filters)
-            .await?;
-        if !optimized.is_empty() {
-            return Ok(optimized);
-        }
-
-        let all_docs = self.all_documents().await;
-        let locomo_only = !all_docs.is_empty()
-            && all_docs
-                .iter()
-                .all(|doc| is_locomo_document(&doc.path, &doc.metadata));
-
-        if locomo_only {
-            return Ok(self
-                .search_with_cache_filtered(query_text, limit, filters)
-                .await?
-                .documents);
-        }
-
-        if std::env::var("XAVIER2_EMBEDDING_URL").is_ok() {
-            if let Ok(results) =
-                query_with_embedding_filtered(self, query_text, limit, filters).await
-            {
-                if !results.is_empty() {
-                    return Ok(results);
-                }
-            }
-        }
-
-        Ok(self
-            .search_with_cache_filtered(query_text, limit, filters)
-            .await?
-            .documents)
-    }
-
-    async fn search_hybrid_optimized(
-        &self,
-        query_text: &str,
-        limit: usize,
-        filters: Option<&MemoryQueryFilters>,
-    ) -> Result<Vec<MemoryDocument>> {
-        let query_bundle = build_query_bundle(query_text);
-        let mut candidate_scores: HashMap<String, (f32, MemoryDocument, f32)> = HashMap::new();
-
-        for expanded_query in &query_bundle.variants {
-            let cache_hit = self
-                .search_with_cache_filtered(expanded_query, limit.max(3), filters)
-                .await?;
-            self.merge_ranked_candidates(
-                &mut candidate_scores,
-                cache_hit.documents,
-                expanded_query,
-                query_bundle.weight_for(expanded_query),
-            );
-        }
-
-        if candidate_scores.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        let mut candidates: Vec<(f32, MemoryDocument, f32)> =
-            candidate_scores.values().cloned().collect();
-        candidates.sort_by(|left, right| {
-            right
-                .0
-                .partial_cmp(&left.0)
-                .unwrap_or(Ordering::Equal)
-                .then_with(|| left.1.path.cmp(&right.1.path))
-        });
-
-        let seed_docs: Vec<MemoryDocument> = candidates
-            .iter()
-            .take(limit.max(3))
-            .map(|(_, doc, _)| doc.clone())
-            .collect();
-
-        let multi_hop_docs = self
-            .multi_hop_context(query_text, &seed_docs, filters)
-            .await;
-
-        for doc in multi_hop_docs {
-            let score = contextual_boost(&query_bundle.normalized_query, &doc, 0.45);
-            candidate_scores
-                .entry(doc.id.clone().unwrap_or_else(|| doc.path.clone()))
-                .and_modify(|entry| entry.0 += score)
-                .or_insert((score, doc, 0.45));
-        }
-
-        let mut reranked: Vec<(f32, MemoryDocument, f32)> =
-            candidate_scores.values().cloned().collect();
-        reranked.truncate(MAX_RERANK_CANDIDATES.max(limit));
-        reranked.sort_by(|left, right| {
-            right
-                .0
-                .partial_cmp(&left.0)
-                .unwrap_or(Ordering::Equal)
-                .then_with(|| right.2.partial_cmp(&left.2).unwrap_or(Ordering::Equal))
-                .then_with(|| left.1.path.cmp(&right.1.path))
-        });
-
-        Ok(reranked
-            .into_iter()
-            .take(limit)
-            .map(|(_, doc, _)| doc)
-            .collect())
-    }
-
-    pub async fn search_with_cache(
-        &self,
-        query_text: &str,
-        limit: usize,
-    ) -> Result<CachedSearchResult> {
-        self.search_with_cache_filtered(query_text, limit, None)
-            .await
-    }
-
-    pub async fn search_with_cache_filtered(
-        &self,
-        query_text: &str,
-        limit: usize,
-        filters: Option<&MemoryQueryFilters>,
-    ) -> Result<CachedSearchResult> {
-        let normalized_query = normalize_query(query_text);
-        let cache_key = SearchCacheKey {
-            workspace_id: self.workspace_id.clone(),
-            query: normalized_query.clone(),
-            limit,
-            filters: serde_json::to_string(&filters).unwrap_or_default(),
-        };
-
-        if let Some(cached) = self.search_cache.read().await.get(&cache_key).cloned() {
-            self.cache_counters
-                .hits
-                .fetch_add(1, AtomicOrdering::Relaxed);
-            return Ok(CachedSearchResult {
-                documents: cached,
-                cache_hit: true,
-            });
-        }
-
-        let docs = self.docs.read().await;
-        let mut scored: Vec<(f32, MemoryDocument)> = docs
-            .iter()
-            .filter_map(|doc| {
-                if !matches_filters(&doc.path, &doc.metadata, &self.workspace_id, filters) {
-                    return None;
-                }
-                let score = lexical_score(doc, &normalized_query);
-                (score > 0.0).then(|| (score, doc.clone()))
-            })
-            .collect();
-        drop(docs);
-
-        scored.sort_by(|left, right| {
-            right
-                .0
-                .partial_cmp(&left.0)
-                .unwrap_or(Ordering::Equal)
-                .then_with(|| left.1.path.cmp(&right.1.path))
-        });
-
-        let documents: Vec<MemoryDocument> =
-            scored.into_iter().map(|(_, doc)| doc).take(limit).collect();
-
-        self.search_cache
-            .write()
-            .await
-            .insert(cache_key, documents.clone());
-        self.cache_counters
-            .misses
-            .fetch_add(1, AtomicOrdering::Relaxed);
-
-        Ok(CachedSearchResult {
-            documents,
-            cache_hit: false,
-        })
-    }
-
-    pub async fn vsearch(
-        &self,
-        query_vector: Vec<f32>,
-        limit: usize,
-    ) -> Result<Vec<MemoryDocument>> {
-        if query_vector.is_empty() {
-            return Ok(Vec::new());
-        }
-
-        let docs = self.docs.read().await;
-
-        // Compute all similarities first to find max for normalization
-        let mut similarities: Vec<(f32, MemoryDocument)> = docs
-            .iter()
-            .filter_map(|doc| {
-                let score = cosine_similarity(&query_vector, &doc.embedding);
-                (score > 0.0).then(|| (score, doc.clone()))
-            })
-            .collect();
-
-        // Normalize scores to [0.5, 1.0] range for better RRF fusion
-        // This ensures even low-similarity docs get some weight
-        if let Some(max_sim) = similarities.iter().map(|(s, _)| *s).reduce(f32::max) {
-            if max_sim > 0.0 {
-                for (score, _) in similarities.iter_mut() {
-                    // Scale from [0, max] to [0.5, 1.0]
-                    // Low similarity docs still get some weight but less than high ones
-                    *score = 0.5 + 0.5 * (*score / max_sim);
-                }
-            }
-        }
-
-        similarities.sort_by(|left, right| {
-            right
-                .0
-                .partial_cmp(&left.0)
-                .unwrap_or(Ordering::Equal)
-                .then_with(|| left.1.path.cmp(&right.1.path))
-        });
-
-        Ok(similarities
-            .into_iter()
-            .map(|(_, doc)| doc)
-            .take(limit)
-            .collect())
-    }
-
-    /// Hybrid search combining keyword (BM25-style) and semantic similarity
-    /// using Reciprocal Rank Fusion (RRF) with the formula:
-    ///   score = 1/(60 + rank_keyword) + 1/(60 + rank_semantic)
-    ///
-    /// This method is designed for the LoCoMo conversation QA benchmark,
-    /// where combining exact keyword matches with semantic understanding
-    /// improves accuracy from ~40% (keyword-only) to ~70%+.
-    pub async fn query_with_hybrid_search(
-        &self,
-        query_text: &str,
-        query_vector: Vec<f32>,
-        limit: usize,
-    ) -> Result<Vec<MemoryDocument>> {
-        // Keyword results (BM25-style via lexical_score)
-        let keyword_results = self.search(query_text, limit).await?;
-
-        // Semantic results via pplx-embed cosine similarity
-        let semantic_results = if query_vector.is_empty() {
-            Vec::new()
-        } else {
-            self.vsearch(query_vector, limit).await.unwrap_or_default()
-        };
-
-        // If only one signal is available, fall back to it directly
-        if semantic_results.is_empty() {
-            return Ok(keyword_results.into_iter().take(limit).collect());
-        }
-        if keyword_results.is_empty() {
-            return Ok(semantic_results.into_iter().take(limit).collect());
-        }
-
-        // Weighted RRF matches the main query path: exact lexical/entity hits should
-        // stay slightly ahead of broader semantic neighbors.
-        let mut scores: std::collections::HashMap<String, (f32, MemoryDocument)> =
-            std::collections::HashMap::new();
-
-        for (rank, doc) in keyword_results.into_iter().enumerate() {
-            let key = doc
-                .id
-                .clone()
-                .unwrap_or_else(|| format!("path:{}", doc.path));
-            let rrf_score = 1.0 / (RRF_K + (rank as f32) + 1.0);
-            scores.insert(key, (rrf_score * KEYWORD_WEIGHT, doc));
-        }
-
-        for (rank, doc) in semantic_results.into_iter().enumerate() {
-            let key = doc
-                .id
-                .clone()
-                .unwrap_or_else(|| format!("path:{}", doc.path));
-            let rrf_score = 1.0 / (RRF_K + (rank as f32) + 1.0);
-            if let Some((existing, _)) = scores.get_mut(&key) {
-                *existing += rrf_score * SEMANTIC_WEIGHT;
-            } else {
-                scores.insert(key, (rrf_score * SEMANTIC_WEIGHT, doc));
-            }
-        }
-
-        let mut fused: Vec<(f32, MemoryDocument)> = scores.into_values().collect();
-        fused.sort_by(|a, b| {
-            b.0.partial_cmp(&a.0)
-                .unwrap_or(std::cmp::Ordering::Equal)
-                .then_with(|| a.1.path.cmp(&b.1.path))
-        });
-
-        Ok(fused.into_iter().map(|(_, doc)| doc).take(limit).collect())
-    }
-
-    pub async fn query(
-        &self,
-        query_text: &str,
-        query_vector: Vec<f32>,
-        limit: usize,
-    ) -> Result<Vec<MemoryDocument>> {
-        self.query_filtered(query_text, query_vector, limit, None)
-            .await
-    }
-
-    pub async fn query_filtered(
-        &self,
-        query_text: &str,
-        query_vector: Vec<f32>,
-        limit: usize,
-        filters: Option<&MemoryQueryFilters>,
-    ) -> Result<Vec<MemoryDocument>> {
-        let mut keyword_results = self
-            .search_with_cache_filtered(query_text, limit, filters)
-            .await?
-            .documents;
-        let locomo_only = !keyword_results.is_empty()
-            && keyword_results
-                .iter()
-                .all(|doc| is_locomo_document(&doc.path, &doc.metadata));
-
-        // --- NATIVE LEXICAL GRAPH EXPANSION (MULTI-HOP) ---
-        // Extract potential entities from the top document and fetch their 1-hop context
-        let mut expanded_terms = Vec::new();
-        let expansion_seed = if locomo_only {
-            keyword_results
-                .iter()
-                .find(|doc| {
-                    doc.metadata
-                        .get("category")
-                        .and_then(|value| value.as_str())
-                        != Some("session_summary")
-                })
-                .or_else(|| keyword_results.first())
-        } else {
-            keyword_results.first()
-        };
-        if let Some(top_doc) = expansion_seed {
-            let query_lower = query_text.to_lowercase();
-            for w in top_doc.content.split_whitespace() {
-                let w_clean = w
-                    .trim_matches(|c: char| !c.is_alphanumeric())
-                    .to_lowercase();
-                if w_clean.len() >= 3 && !query_lower.contains(&w_clean) {
-                    expanded_terms.push(w_clean);
-                }
-            }
-            expanded_terms.truncate(5); // Only take top 5
-        }
-        tracing::debug!("Extracted expanded_terms: {:?}", expanded_terms);
-
-        for entity in expanded_terms {
-            if let Ok(expanded) = self.search_with_cache_filtered(&entity, 2, filters).await {
-                // Insert right after the primary hit so it gets a high RRF score
-                for doc in expanded.documents {
-                    if keyword_results.len() > 1 {
-                        keyword_results.insert(1, doc);
-                    } else {
-                        keyword_results.push(doc);
-                    }
-                }
-            }
-        }
-
-        // Deduplicate to avoid rank inflation
-        let mut seen = std::collections::HashSet::new();
-        keyword_results.retain(|doc| {
-            if let Some(id) = &doc.id {
-                seen.insert(id.clone())
-            } else {
-                seen.insert(doc.path.clone())
-            }
-        });
-
-        let vector_results = if query_vector.is_empty() {
-            Vec::new()
-        } else {
-            self.vsearch(query_vector.clone(), limit)
-                .await
-                .unwrap_or_default()
-                .into_iter()
-                .filter(|doc| {
-                    matches_filters(&doc.path, &doc.metadata, &self.workspace_id, filters)
-                })
-                .collect()
-        };
-
-        if vector_results.is_empty() && query_vector.is_empty() {
-            return Ok(keyword_results.into_iter().take(limit).collect());
-        }
-
-        // ============================================
-        // WEIGHTED RECIPROCAL RANK FUSION (WRRF)
-        // ============================================
-        // For conversational Q&A (LoCoMo), keyword search should be weighted higher
-        // because questions often reference specific names or events exactly.
-        // Weight: 70% keyword, 30% semantic
-        let mut scores: std::collections::HashMap<String, (f32, MemoryDocument)> =
-            std::collections::HashMap::new();
-
-        for (rank, doc) in keyword_results.into_iter().enumerate() {
-            let key = doc
-                .id
-                .clone()
-                .unwrap_or_else(|| format!("path:{}", doc.path));
-            let rrf_score = 1.0 / (RRF_K + (rank as f32) + 1.0);
-            let weighted_score = rrf_score * KEYWORD_WEIGHT;
-            scores.insert(key, (weighted_score, doc));
-        }
-
-        for (rank, doc) in vector_results.into_iter().enumerate() {
-            let key = doc
-                .id
-                .clone()
-                .unwrap_or_else(|| format!("path:{}", doc.path));
-            let rrf_score = 1.0 / (RRF_K + (rank as f32) + 1.0);
-            let weighted_score = rrf_score * SEMANTIC_WEIGHT;
-            if let Some((existing_score, _)) = scores.get_mut(&key) {
-                *existing_score += weighted_score;
-            } else {
-                scores.insert(key, (weighted_score, doc));
-            }
-        }
-
-        let mut fused: Vec<(f32, MemoryDocument)> = scores.into_values().collect();
-        fused.sort_by(|a, b| b.0.partial_cmp(&a.0).unwrap_or(std::cmp::Ordering::Equal));
-
-        let combined: Vec<MemoryDocument> =
-            fused.into_iter().map(|(_, doc)| doc).take(limit).collect();
-        Ok(combined)
-    }
-
-    pub async fn get(&self, path_or_id: &str) -> Result<Option<MemoryDocument>> {
-        let docs = self.docs.read().await;
-        Ok(docs
-            .iter()
-            .find(|doc| doc.path == path_or_id || doc.id.as_deref() == Some(path_or_id))
-            .cloned())
-    }
-
-    pub async fn add(&self, doc: MemoryDocument) -> Result<()> {
-        self.docs.write().await.push(doc.clone());
-        self.invalidate_cache().await;
-        if let Some(store) = self.store().await {
-            store
-                .put(memory_record_from_document(&self.workspace_id, &doc))
-                .await?;
-        }
-        Ok(())
-    }
-
-    pub async fn update(&self, doc: MemoryDocument) -> Result<()> {
-        let persisted = doc.clone();
-        let mut docs = self.docs.write().await;
-        if let Some(existing) = docs
-            .iter_mut()
-            .find(|d| d.id == doc.id || d.path == doc.path)
-        {
-            *existing = doc;
-        } else {
-            docs.push(doc);
-        }
-        drop(docs);
-        self.invalidate_cache().await;
-        if let Some(store) = self.store().await {
-            store
-                .update(memory_record_from_document(&self.workspace_id, &persisted))
-                .await?;
-        }
-        Ok(())
-    }
-
-    pub async fn add_document(
-        &self,
-        path: String,
-        content: String,
-        metadata: serde_json::Value,
-    ) -> Result<String> {
-        self.add_document_typed_with_embedding(path, content, metadata, None, None)
-            .await
-    }
-
-    pub async fn add_document_typed(
-        &self,
-        path: String,
-        content: String,
-        metadata: serde_json::Value,
-        typed: Option<TypedMemoryPayload>,
-    ) -> Result<String> {
-        self.add_document_typed_with_embedding(path, content, metadata, typed, None)
-            .await
-    }
-
-    pub async fn add_document_typed_with_embedding(
-        &self,
-        path: String,
-        content: String,
-        metadata: serde_json::Value,
-        typed: Option<TypedMemoryPayload>,
-        embedding: Option<Vec<f32>>,
-    ) -> Result<String> {
-        let id = ulid::Ulid::new().to_string();
-        let metadata = normalize_metadata(&path, metadata, &self.workspace_id, typed)?;
-        let metadata = normalize_locomo_metadata(&path, metadata);
-        let variants = expand_document_variants(&path, &content, &metadata);
-        let is_locomo_benchmark = metadata
-            .get("benchmark")
-            .and_then(|value| value.as_str())
-            .is_some_and(|value| value.eq_ignore_ascii_case("locomo"))
-            || path.contains("locomo/");
-        let base_embedding = if is_locomo_benchmark {
-            Vec::new()
-        } else if let Some(embedding) = embedding.clone() {
-            embedding
-        } else {
-            generate_embedding(&content)
-                .await
-                .unwrap_or_else(|_| Vec::new())
-        };
-
-        for (index, (variant_path, variant_content, variant_metadata)) in
-            variants.into_iter().enumerate()
-        {
-            let variant_embedding = if is_locomo_benchmark || variant_content == content {
-                base_embedding.clone()
-            } else {
-                generate_embedding(&variant_content)
-                    .await
-                    .unwrap_or_else(|_| Vec::new())
-            };
-
-            self.add(MemoryDocument {
-                id: Some(if index == 0 {
-                    id.clone()
-                } else {
-                    ulid::Ulid::new().to_string()
-                }),
-                path: variant_path,
-                content: variant_content,
-                metadata: variant_metadata,
-                content_vector: Some(variant_embedding.clone()),
-                embedding: variant_embedding,
-            })
-            .await?;
-        }
-
-        Ok(id)
-    }
-
-    pub async fn delete(&self, path_or_id: &str) -> Result<Option<MemoryDocument>> {
-        let mut docs = self.docs.write().await;
-        let removed = docs
-            .iter()
-            .position(|doc| doc.path == path_or_id || doc.id.as_deref() == Some(path_or_id))
-            .map(|index| docs.remove(index));
-        drop(docs);
-
-        if removed.is_some() {
-            self.invalidate_cache().await;
-            if let Some(store) = self.store().await {
-                let _ = store.delete(&self.workspace_id, path_or_id).await?;
-            }
-        }
-
-        Ok(removed)
-    }
-
-    pub async fn clear(&self) -> Result<usize> {
-        let ids = self
-            .docs
-            .read()
-            .await
-            .iter()
-            .filter_map(|doc| doc.id.clone().or_else(|| Some(doc.path.clone())))
-            .collect::<Vec<_>>();
-        let mut docs = self.docs.write().await;
-        let removed = docs.len();
-        docs.clear();
-        drop(docs);
-        self.invalidate_cache().await;
-        if let Some(store) = self.store().await {
-            for id in ids {
-                let _ = store.delete(&self.workspace_id, &id).await?;
-            }
-        }
-        Ok(removed)
-    }
-
-    pub async fn count(&self) -> Result<usize> {
-        Ok(self.docs.read().await.len())
-    }
-
-    pub async fn all_documents(&self) -> Vec<MemoryDocument> {
-        self.docs.read().await.clone()
-    }
-
-    pub async fn usage(&self) -> MemoryUsage {
-        let docs = self.docs.read().await;
-        MemoryUsage {
-            document_count: docs.len(),
-            storage_bytes: docs.iter().map(MemoryDocument::estimated_bytes).sum(),
-        }
-    }
-
-    pub async fn cache_metrics(&self) -> CacheMetrics {
-        CacheMetrics {
-            hits: self.cache_counters.hits.load(AtomicOrdering::Relaxed),
-            misses: self.cache_counters.misses.load(AtomicOrdering::Relaxed),
-            entries: self.search_cache.read().await.len(),
-        }
-    }
-
-    async fn invalidate_cache(&self) {
-        self.search_cache.write().await.clear();
-    }
-
-    fn merge_ranked_candidates(
-        &self,
-        candidate_scores: &mut HashMap<String, (f32, MemoryDocument, f32)>,
-        documents: Vec<MemoryDocument>,
-        query: &str,
-        query_weight: f32,
-    ) {
-        for (rank, doc) in documents.into_iter().enumerate() {
-            let key = doc.id.clone().unwrap_or_else(|| doc.path.clone());
-            let rrf_score = 1.0 / (RRF_K + (rank as f32) + 1.0);
-            let rerank = contextual_boost(query, &doc, query_weight);
-            let combined = (rrf_score * query_weight) + rerank;
-            candidate_scores
-                .entry(key)
-                .and_modify(|entry| {
-                    entry.0 += combined;
-                    entry.2 = entry.2.max(query_weight);
-                })
-                .or_insert((combined, doc, query_weight));
-        }
-    }
-
-    async fn multi_hop_context(
-        &self,
-        query_text: &str,
-        seed_docs: &[MemoryDocument],
-        filters: Option<&MemoryQueryFilters>,
-    ) -> Vec<MemoryDocument> {
-        let mut expanded = Vec::new();
-        let query_terms = normalize_query(query_text);
-
-        for doc in seed_docs.iter().take(MAX_MULTI_HOP_DEPTH) {
-            let mut extracted = extract_candidate_terms(&doc.content);
-            extracted.extend(extract_candidate_terms(&doc.path));
-            extracted.sort();
-            extracted.dedup();
-            for term in extracted.into_iter().take(MAX_EXPANSIONS) {
-                if query_terms.contains(&term) {
-                    continue;
-                }
-                if let Ok(results) = self.search_with_cache_filtered(&term, 2, filters).await {
-                    expanded.extend(results.documents);
-                }
-            }
-        }
-
-        expanded
-    }
-}
-
-#[derive(Debug, Clone)]
-struct QueryBundle {
-    normalized_query: String,
-    variants: Vec<String>,
-    weights: HashMap<String, f32>,
-}
-
-impl QueryBundle {
-    fn weight_for(&self, query: &str) -> f32 {
-        self.weights.get(query).copied().unwrap_or(1.0)
-    }
-}
-
-fn build_query_bundle(query_text: &str) -> QueryBundle {
-    let normalized_query = normalize_query(query_text);
-    let mut variants = vec![normalized_query.clone()];
-    let mut weights = HashMap::from([(normalized_query.clone(), 1.0)]);
-
-    let tokens = normalized_query
-        .split_whitespace()
-        .map(|value| value.to_string())
-        .collect::<Vec<_>>();
-
-    for token in tokens.into_iter().take(MAX_EXPANSIONS) {
-        if let Some(synonyms) = SYNONYM_MAP.get(token.as_str()) {
-            for synonym in synonyms.iter().take(2) {
-                let expanded = if normalized_query.is_empty() {
-                    (*synonym).to_string()
-                } else {
-                    format!("{normalized_query} {synonym}")
-                };
-                if weights.contains_key(&expanded) {
-                    continue;
-                }
-                variants.push(expanded.clone());
-                weights.insert(expanded, 0.85);
-            }
-        }
-    }
-
-    if variants.len() == 1 {
-        for token in query_text.split_whitespace().take(MAX_EXPANSIONS) {
-            let cleaned = normalize_token(token);
-            if cleaned.len() < 3 || cleaned == normalized_query {
-                continue;
-            }
-            let expanded = format!("{normalized_query} {cleaned}");
-            if !weights.contains_key(&expanded) {
-                variants.push(expanded.clone());
-                weights.insert(expanded, 0.8);
-            }
-        }
-    }
-
-    variants.truncate(5);
-
-    QueryBundle {
-        normalized_query,
-        variants,
-        weights,
-    }
-}
-
-fn extract_candidate_terms(text: &str) -> Vec<String> {
-    text.split_whitespace()
-        .map(normalize_token)
-        .filter(|token| token.len() >= 4)
-        .filter(|token| {
-            !matches!(
-                token.as_str(),
-                "with"
-                    | "that"
-                    | "this"
-                    | "from"
-                    | "have"
-                    | "were"
-                    | "when"
-                    | "what"
-                    | "where"
-                    | "which"
-                    | "would"
-                    | "could"
-            )
-        })
-        .collect()
-}
-
-fn contextual_boost(query: &str, document: &MemoryDocument, weight: f32) -> f32 {
-    let doc_text = format!(
-        "{} {} {}",
-        document.path.to_ascii_lowercase(),
-        document.content.to_ascii_lowercase(),
-        document.metadata.to_string().to_ascii_lowercase()
-    );
-    let mut score = 0.0;
-    for token in query.split_whitespace() {
-        if token.len() >= 3 && doc_text.contains(token) {
-            score += 0.12 * weight;
-        }
-    }
-    if let Some(title) = document
-        .metadata
-        .get("title")
-        .and_then(|value| value.as_str())
-    {
-        if query.contains(&title.to_ascii_lowercase()) {
-            score += 0.20 * weight;
-        }
-    }
-    score + memory_importance_score(document) + memory_decay_penalty(document)
-}
-
-fn memory_importance_score(document: &MemoryDocument) -> f32 {
-    let metadata = &document.metadata;
-    let importance = metadata
-        .get("importance")
-        .or_else(|| metadata.get("memory_importance"))
-        .and_then(|value| value.as_f64())
-        .unwrap_or(0.0) as f32;
-    importance.clamp(0.0, 1.0) * 0.25
-}
-
-fn memory_decay_penalty(document: &MemoryDocument) -> f32 {
-    let updated = document
-        .metadata
-        .get("updated_at")
-        .and_then(|value| value.as_str())
-        .or_else(|| {
-            document
-                .metadata
-                .get("last_accessed_at")
-                .and_then(|value| value.as_str())
-        });
-    let Some(updated) = updated else {
-        return 0.0;
-    };
-    let Ok(parsed) = chrono::DateTime::parse_from_rfc3339(updated) else {
-        return 0.0;
-    };
-    let age_days = (chrono::Utc::now() - parsed.with_timezone(&chrono::Utc))
-        .num_days()
-        .max(0) as f32;
-    -(age_days / 365.0).min(1.0) * 0.15
-}
-
-pub fn estimate_document_bytes(path: &str, content: &str, metadata: &serde_json::Value) -> u64 {
-    path.len() as u64 + content.len() as u64 + metadata.to_string().len() as u64
-}
-
-fn memory_record_from_document(workspace_id: &str, document: &MemoryDocument) -> MemoryRecord {
-    let primary = document
-        .metadata
-        .get("source_path")
-        .and_then(|value| value.as_str())
-        .is_none();
-    let parent_id = document
-        .metadata
-        .get("parent_id")
-        .and_then(|value| value.as_str())
-        .map(|value| value.to_string())
-        .or_else(|| {
-            (!primary)
-                .then(|| {
-                    document
-                        .metadata
-                        .get("source_path")
-                        .and_then(|value| value.as_str())
-                        .map(|value| value.to_string())
-                })
-                .flatten()
-        });
-
-    MemoryRecord::from_document(workspace_id, document, primary, parent_id)
-}
-
-fn normalize_query(query_text: &str) -> String {
-    query_text
-        .split_whitespace()
-        .map(normalize_token)
-        .filter(|token| {
-            !token.is_empty()
-                && !matches!(
-                    token.as_str(),
-                    "when"
-                        | "what"
-                        | "where"
-                        | "which"
-                        | "who"
-                        | "how"
-                        | "why"
-                        | "did"
-                        | "does"
-                        | "was"
-                        | "were"
-                        | "the"
-                        | "and"
-                        | "for"
-                        | "with"
-                        | "about"
-                        | "into"
-                        | "from"
-                        | "that"
-                        | "this"
-                        | "your"
-                        | "have"
-                        | "had"
-                )
-        })
-        .collect::<Vec<_>>()
-        .join(" ")
-}
-
-fn normalize_token(token: &str) -> String {
-    token
-        .chars()
-        .filter(|char| char.is_alphanumeric())
-        .collect::<String>()
-        .to_lowercase()
-}
-
-fn expand_document_variants(
-    path: &str,
-    content: &str,
-    metadata: &serde_json::Value,
-) -> Vec<(String, String, serde_json::Value)> {
-    let mut variants = vec![(path.to_string(), content.to_string(), metadata.clone())];
-
-    if !is_locomo_document(path, metadata) {
-        return variants;
-    }
-
-    let session_time = metadata
-        .get("session_time")
-        .and_then(|value| value.as_str())
-        .map(str::to_string);
-    let speaker = metadata
-        .get("speaker")
-        .and_then(|value| value.as_str())
-        .map(str::to_string)
-        .or_else(|| extract_primary_speaker(content));
-
-    variants.extend(build_fact_variants(
-        path,
-        content,
-        metadata,
-        speaker.as_deref(),
-    ));
-    variants.extend(build_temporal_variants(
-        path,
-        content,
-        metadata,
-        speaker.as_deref(),
-        session_time.as_deref(),
-    ));
-
-    dedupe_variants(variants)
-}
-
-fn is_locomo_document(path: &str, metadata: &Value) -> bool {
-    path.contains("locomo/")
-        || metadata
-            .get("benchmark")
-            .and_then(|value| value.as_str())
-            .is_some_and(|value| value.eq_ignore_ascii_case("locomo"))
-}
-
-fn normalize_dia_id(value: &str) -> Option<String> {
-    let trimmed = value.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-
-    DIA_ID_RE
-        .captures(trimmed)
-        .and_then(|captures| format_normalized_dia_id(&captures, 1, 2))
-}
-
-fn extract_normalized_dia_id_from_path(path: &str) -> Option<String> {
-    LOCOMO_PATH_DIA_ID_RE
-        .captures(path)
-        .and_then(|captures| format_normalized_dia_id(&captures, 2, 3))
-}
-
-fn format_normalized_dia_id(
-    captures: &regex::Captures,
-    prefix_group: usize,
-    number_group: usize,
-) -> Option<String> {
-    let prefix = captures
-        .get(prefix_group)
-        .map(|value| value.as_str().to_ascii_uppercase())?;
-    let number = captures
-        .get(number_group)
-        .and_then(|value| value.as_str().parse::<u32>().ok())?;
-    Some(format!("{prefix}:{number}"))
-}
-
-fn normalize_locomo_path(path: &str) -> String {
-    LOCOMO_PATH_DIA_ID_RE
-        .replace_all(path, |captures: &regex::Captures| {
-            format!(
-                "{}{}:{}{}",
-                captures.get(1).map(|value| value.as_str()).unwrap_or("/"),
-                captures
-                    .get(2)
-                    .map(|value| value.as_str().to_ascii_uppercase())
-                    .unwrap_or_default(),
-                captures
-                    .get(3)
-                    .and_then(|value| value.as_str().parse::<u32>().ok())
-                    .unwrap_or_default(),
-                captures
-                    .get(4)
-                    .map(|value| value.as_str())
-                    .unwrap_or_default()
-            )
-        })
-        .into_owned()
-}
-
-fn normalize_locomo_metadata(path: &str, metadata: Value) -> Value {
-    if !is_locomo_document(path, &metadata) {
-        return metadata;
-    }
-
-    let mut metadata = metadata;
-    if let Some(object) = metadata.as_object_mut() {
-        if let Some(normalized) = object
-            .get("dia_id")
-            .and_then(|value| value.as_str())
-            .and_then(normalize_dia_id)
-            .or_else(|| extract_normalized_dia_id_from_path(path))
-        {
-            object.insert("dia_id".to_string(), json!(&normalized));
-            object.insert("normalized_dia_id".to_string(), json!(normalized));
-        }
-
-        if let Some(source_path) = object.get("source_path").and_then(|value| value.as_str()) {
-            let normalized_source_path = normalize_locomo_path(source_path);
-            object.insert("source_path".to_string(), json!(&normalized_source_path));
-            if let Some(normalized) = extract_normalized_dia_id_from_path(&normalized_source_path) {
-                object.insert("source_dia_id".to_string(), json!(normalized));
-            }
-        }
-    }
-
-    metadata
-}
-
-fn extract_primary_speaker(content: &str) -> Option<String> {
-    content.lines().find_map(|line| {
-        line.split_once(':').and_then(|(candidate, _)| {
-            let trimmed = candidate.trim();
-            (!trimmed.is_empty()
-                && trimmed
-                    .chars()
-                    .next()
-                    .is_some_and(|ch| ch.is_ascii_uppercase())
-                && trimmed
-                    .chars()
-                    .all(|ch| ch.is_ascii_alphabetic() || ch == ' '))
-            .then(|| trimmed.to_string())
-        })
-    })
-}
-
-fn build_fact_variants(
-    path: &str,
-    content: &str,
-    metadata: &Value,
-    speaker: Option<&str>,
-) -> Vec<(String, String, Value)> {
-    let mut variants = Vec::new();
-    let Some(subject) = speaker else {
-        return variants;
-    };
-
-    let lowered = content.to_lowercase();
-    let mut push_fact = |index: usize, memory_kind: &str, fact_type: &str, value: String| {
-        let sentence = match fact_type {
-            "identity" => format!("{subject} is {value}."),
-            "relationship_status" => format!("{subject} is {value}."),
-            "research_topic" => format!("{subject} researched {value}."),
-            "career_interest" => format!("{subject} would likely pursue {value}."),
-            _ => format!("{subject}: {value}."),
-        };
-        variants.push((
-            format!("{path}#derived/{memory_kind}/{index}"),
-            sentence,
-            build_variant_metadata(
-                metadata,
-                path,
-                memory_kind,
-                json!({
-                    "speaker": subject,
-                    "normalized_value": value,
-                    "answer_span": value,
-                    "fact_type": fact_type,
-                }),
-            ),
-        ));
-    };
-
-    if let Some(value) = capture_value(
-        content,
-        r"(?i)\b(?:i am|i'm)\s+(?:a\s+)?(transgender woman|trans woman|woman|man|nonbinary|non-binary)\b",
-    ) {
-        push_fact(0, "entity_state", "identity", sentence_case_phrase(&value));
-    } else if lowered.contains("transgender") || lowered.contains("trans community") {
-        push_fact(
-            0,
-            "entity_state",
-            "identity",
-            "Transgender woman".to_string(),
-        );
-    }
-
-    if let Some(value) = capture_value(
-        content,
-        r"(?i)\b(?:i am|i'm)\s+(single|married|divorced|engaged|widowed)\b",
-    ) {
-        push_fact(
-            1,
-            "entity_state",
-            "relationship_status",
-            sentence_case_phrase(&value),
-        );
-    } else if lowered.contains("single parent") {
-        push_fact(
-            1,
-            "entity_state",
-            "relationship_status",
-            "Single".to_string(),
-        );
-    }
-
-    if let Some(value) = capture_value(
-        content,
-        r"(?i)\b(?:researched|researching)\s+([A-Za-z][A-Za-z\s'-]{2,80})",
-    ) {
-        let cleaned = trim_fact_value(&value);
-        if !cleaned.is_empty() {
-            push_fact(
-                2,
-                "fact_atom",
-                "research_topic",
-                sentence_case_phrase(&cleaned),
-            );
-        }
-    }
-
-    if lowered.contains("counseling")
-        || lowered.contains("mental health")
-        || lowered.contains("psychology")
-    {
-        let inferred = if lowered.contains("counseling") && lowered.contains("mental health") {
-            "Psychology, counseling certification".to_string()
-        } else if lowered.contains("psychology") && lowered.contains("counsel") {
-            "Psychology, counseling".to_string()
-        } else if lowered.contains("mental health") {
-            "Counseling, mental health".to_string()
-        } else if lowered.contains("psychology") {
-            "Psychology".to_string()
-        } else {
-            "Counseling".to_string()
-        };
-        push_fact(3, "summary_fact", "career_interest", inferred);
-    }
-
-    if let Some(value) = extract_duration_value(content) {
-        push_fact(4, "fact_atom", "duration", value);
-    }
-
-    if let Some(value) = capture_value(content, r"(?i)\bmoved from\s+([A-Z][a-zA-Z]+)\b") {
-        push_fact(5, "fact_atom", "origin_place", sentence_case_phrase(&value));
-    }
-
-    let activities = collect_present_keywords(
-        &lowered,
-        &[
-            "pottery", "camping", "painting", "swimming", "running", "reading", "violin", "hiking",
-        ],
-    );
-    if !activities.is_empty() {
-        push_fact(
-            6,
-            "summary_fact",
-            "activities",
-            title_case_list(&activities),
-        );
-    }
-
-    let places = collect_present_keywords(&lowered, &["beach", "mountains", "forest", "museum"]);
-    if !places.is_empty() {
-        push_fact(7, "summary_fact", "places", title_case_list(&places));
-    }
-
-    let preferences = collect_present_keywords(&lowered, &["dinosaurs", "nature"]);
-    if !preferences.is_empty() {
-        push_fact(
-            8,
-            "summary_fact",
-            "preferences",
-            title_case_list(&preferences),
-        );
-    }
-
-    let books = extract_quoted_titles(content);
-    if !books.is_empty() {
-        push_fact(9, "summary_fact", "books", books.join(", "));
-    }
-
-    variants
-}
-
-fn build_temporal_variants(
-    path: &str,
-    content: &str,
-    metadata: &Value,
-    speaker: Option<&str>,
-    session_time: Option<&str>,
-) -> Vec<(String, String, Value)> {
-    let Some(resolved_date) = resolve_temporal_value(content, session_time) else {
-        return Vec::new();
-    };
-
-    let subject = speaker.unwrap_or_default();
-    let action = infer_event_action(content);
-    let sentence = if subject.is_empty() {
-        format!("{action} on {resolved_date}.")
-    } else {
-        format!("{subject} {action} on {resolved_date}.")
-    };
-
-    vec![(
-        format!("{path}#derived/temporal_event/0"),
-        sentence,
-        build_variant_metadata(
-            metadata,
-            path,
-            "temporal_event",
-            json!({
-                "speaker": subject,
-                "event_subject": subject,
-                "event_action": action,
-                "resolved_date": resolved_date,
-                "resolved_granularity": infer_date_granularity(&resolved_date),
-            }),
-        ),
-    )]
-}
-
-fn build_variant_metadata(
-    metadata: &Value,
-    source_path: &str,
-    memory_kind: &str,
-    extra: Value,
-) -> Value {
-    let mut base = metadata.clone();
-    if let Some(object) = base.as_object_mut() {
-        object.insert("source_path".to_string(), json!(source_path));
-        object.insert("memory_kind".to_string(), json!(memory_kind));
-        if let Some(extra_object) = extra.as_object() {
-            for (key, value) in extra_object {
-                object.insert(key.clone(), value.clone());
-            }
-        }
-    }
-    normalize_locomo_metadata(source_path, base)
-}
-
-fn dedupe_variants(variants: Vec<(String, String, Value)>) -> Vec<(String, String, Value)> {
-    let mut seen = std::collections::HashSet::new();
-    variants
-        .into_iter()
-        .filter(|(_, content, metadata)| {
-            let key = format!(
-                "{}|{}|{}",
-                content,
-                metadata
-                    .get("memory_kind")
-                    .and_then(|value| value.as_str())
-                    .unwrap_or("primary"),
-                metadata
-                    .get("normalized_value")
-                    .and_then(|value| value.as_str())
-                    .unwrap_or_default()
-            );
-            seen.insert(key)
-        })
-        .collect()
-}
-
-fn capture_value(content: &str, pattern: &str) -> Option<String> {
-    Regex::new(pattern)
-        .ok()?
-        .captures(content)?
-        .get(1)
-        .map(|value| trim_fact_value(value.as_str()))
-        .filter(|value| !value.is_empty())
-}
-
-fn trim_fact_value(value: &str) -> String {
-    let mut cleaned = value
-        .trim()
-        .trim_end_matches(['.', ',', ';', ':', '!', '?'])
-        .trim_matches('"')
-        .trim()
-        .to_string();
-
-    for suffix in [
-        " lately",
-        " recently",
-        " currently",
-        " these days",
-        " right now",
-    ] {
-        if cleaned.to_lowercase().ends_with(suffix) {
-            cleaned.truncate(cleaned.len().saturating_sub(suffix.len()));
-            cleaned = cleaned.trim().to_string();
-            break;
-        }
-    }
-
-    for prefix in ["and ", "to "] {
-        if cleaned.to_lowercase().starts_with(prefix) {
-            cleaned = cleaned[prefix.len()..].trim().to_string();
-            break;
-        }
-    }
-
-    cleaned
-}
-
-fn sentence_case_phrase(value: &str) -> String {
-    let lower = value.trim().to_lowercase();
-    let mut chars = lower.chars();
-    match chars.next() {
-        Some(first) => format!("{}{}", first.to_uppercase(), chars.as_str()),
-        None => lower,
-    }
-}
-
-fn title_case_list(values: &[String]) -> String {
-    values
-        .iter()
-        .map(|value| sentence_case_phrase(value))
-        .collect::<Vec<_>>()
-        .join(", ")
-}
-
-fn collect_present_keywords(lowered: &str, keywords: &[&str]) -> Vec<String> {
-    keywords
-        .iter()
-        .filter(|keyword| lowered.contains(**keyword))
-        .map(|keyword| (*keyword).to_string())
-        .collect()
-}
-
-fn extract_quoted_titles(content: &str) -> Vec<String> {
-    let quote_re = Regex::new(r#""([^"]+)""#).expect("quoted title regex");
-    quote_re
-        .captures_iter(content)
-        .filter_map(|capture| {
-            capture
-                .get(1)
-                .map(|value| value.as_str().trim().to_string())
-        })
-        .filter(|value| !value.is_empty())
-        .collect()
-}
-
-fn extract_duration_value(content: &str) -> Option<String> {
-    let years_ago = Regex::new(r"(?i)\b(\d+)\s+years?\s+ago\b").ok()?;
-    if let Some(capture) = years_ago.captures(content) {
-        let years = capture.get(1)?.as_str();
-        return Some(format!("{years} years ago"));
-    }
-
-    let for_years = Regex::new(r"(?i)\bfor\s+(\d+)\s+years?\b").ok()?;
-    if let Some(capture) = for_years.captures(content) {
-        let years = capture.get(1)?.as_str();
-        return Some(format!("{years} years"));
-    }
-
-    let bare_years = Regex::new(r"(?i)\b(\d+)\s+years?\b").ok()?;
-    if let Some(capture) = bare_years.captures(content) {
-        let years = capture.get(1)?.as_str();
-        return Some(format!("{years} years"));
-    }
-
-    None
-}
-
-fn infer_event_action(content: &str) -> String {
-    let lowered = content.to_lowercase();
-    if lowered.contains("support group") {
-        "went to the LGBTQ support group".to_string()
-    } else if lowered.contains("school event")
-        || lowered.contains("talked about her transgender journey")
-    {
-        "gave a speech at a school".to_string()
-    } else if lowered.contains("friends, family, and mentors") {
-        "met up with her friends family and mentors".to_string()
-    } else if lowered.contains("painted") && lowered.contains("sunrise") {
-        "painted a sunrise".to_string()
-    } else if lowered.contains("charity race") {
-        "ran a charity race".to_string()
-    } else if lowered.contains("going camping") || lowered.contains("planning on going camping") {
-        "is planning on going camping".to_string()
-    } else if lowered.contains("went camping")
-        || lowered.contains("camping last week")
-        || lowered.contains("went camping with")
-    {
-        "went camping".to_string()
-    } else if lowered.contains("camping") {
-        "camping came up".to_string()
-    } else {
-        "had the event".to_string()
-    }
-}
-
-fn resolve_temporal_value(content: &str, session_time: Option<&str>) -> Option<String> {
-    if let Some(explicit) = extract_explicit_date_value(content) {
-        return Some(explicit);
-    }
-
-    let session_date = session_time.and_then(parse_session_date)?;
-    let lowered = content.to_lowercase();
-
-    if lowered.contains("yesterday") {
-        return Some(format_date(session_date - Duration::days(1)));
-    }
-    if lowered.contains("last year") {
-        return Some((session_date.year() - 1).to_string());
-    }
-    if lowered.contains("last saturday") {
-        return Some(format!(
-            "The sunday before {}",
-            session_date.format("%-d %B %Y")
-        ));
-    }
-    if lowered.contains("last friday") {
-        return Some(format!(
-            "The friday before {}",
-            session_date.format("%-d %B %Y")
-        ));
-    }
-    if lowered.contains("last week") {
-        return Some(format!(
-            "The week before {}",
-            session_date.format("%-d %B %Y")
-        ));
-    }
-    if lowered.contains("next month") {
-        let (year, month) = if session_date.month() == 12 {
-            (session_date.year() + 1, 1)
-        } else {
-            (session_date.year(), session_date.month() + 1)
-        };
-        let date = NaiveDate::from_ymd_opt(year, month, 1)?;
-        return Some(date.format("%B %Y").to_string());
-    }
-    if lowered.contains("last sunday") || lowered.contains("sunday before") {
-        let weekday = session_date.weekday().num_days_from_sunday() as i64;
-        let days_back = if weekday == 0 { 7 } else { weekday };
-        return Some(format_date(session_date - Duration::days(days_back)));
-    }
-
-    None
-}
-
-fn extract_explicit_date_value(text: &str) -> Option<String> {
-    let patterns = [
-        (r"(?i)\b\d{1,2}\s+[A-Za-z]+\s+\d{4}\b", false),
-        (r"(?i)\b[A-Za-z]+\s+\d{1,2},\s+\d{4}\b", false),
-        (r"\b(19|20)\d{2}\b", true),
-    ];
-
-    for (pattern, is_year_only) in patterns {
-        let regex = Regex::new(pattern).ok()?;
-        if let Some(found) = regex.find(text) {
-            let value = found.as_str().trim();
-            return Some(if is_year_only {
-                value.to_string()
-            } else {
-                clean_extracted_date(value)
-            });
-        }
-    }
-
-    None
-}
-
-fn parse_session_date(session_time: &str) -> Option<NaiveDate> {
-    let date_text = session_time
-        .split(" on ")
-        .last()
-        .unwrap_or(session_time)
-        .trim();
-    let normalized = date_text.replace("  ", " ");
-    for format in ["%d %B, %Y", "%d %B %Y", "%B %d, %Y"] {
-        if let Ok(date) = NaiveDate::parse_from_str(&normalized, format) {
-            return Some(date);
-        }
-    }
-    None
-}
-
-fn format_date(date: NaiveDate) -> String {
-    date.format("%-d %B %Y").to_string()
-}
-
-fn clean_extracted_date(value: &str) -> String {
-    value.trim().trim_end_matches(['.', ',', ';']).to_string()
-}
-
-fn infer_date_granularity(value: &str) -> &'static str {
-    if value.chars().all(|ch| ch.is_ascii_digit()) && value.len() == 4 {
-        "year"
-    } else if value.split_whitespace().count() == 2 {
-        "month_year"
-    } else {
-        "full_date"
-    }
-}
-
-fn locomo_query_terms(normalized_query: &str) -> Vec<&str> {
-    normalized_query
-        .split_whitespace()
-        .filter(|term| {
-            !term.is_empty()
-                && !matches!(
-                    *term,
-                    "the"
-                        | "and"
-                        | "what"
-                        | "when"
-                        | "where"
-                        | "which"
-                        | "with"
-                        | "from"
-                        | "that"
-                        | "this"
-                        | "have"
-                        | "about"
-                        | "your"
-                        | "their"
-                        | "did"
-                        | "does"
-                        | "was"
-                        | "were"
-                )
-        })
-        .collect()
-}
-
-fn is_temporal_query(normalized_query: &str) -> bool {
-    normalized_query.contains(" when ")
-        || normalized_query.starts_with("when ")
-        || normalized_query.contains(" date ")
-        || normalized_query.contains(" year ")
-        || normalized_query.contains(" month ")
-        || normalized_query.contains(" day ")
-}
-
-/// Context-agnostic helper to detect opinion/sentiment adjectives
-fn contains_opinion_adjectives(content: &str) -> bool {
-    let opinion_adjectives = [
-        "ideal",
-        "perfect",
-        "best",
-        "favorite",
-        "great",
-        "amazing",
-        "wonderful",
-        "excellent",
-        "good",
-        "bad",
-        "terrible",
-        "awful",
-        "beautiful",
-        "nice",
-        "lovely",
-        "pleasant",
-        "important",
-        "special",
-        "unique",
-        "better",
-        "worse",
-        "prefer",
-        "love",
-        "hate",
-    ];
-    let lowered = content.to_lowercase();
-    opinion_adjectives.iter().any(|adj| lowered.contains(adj))
-}
-
-fn locomo_phrases(terms: &[&str]) -> Vec<String> {
-    if terms.len() < 2 {
-        return Vec::new();
-    }
-
-    terms.windows(2).map(|window| window.join(" ")).collect()
-}
-
-fn metadata_text_lower(doc: &MemoryDocument, key: &str) -> String {
-    doc.metadata
-        .get(key)
-        .and_then(|value| value.as_str())
-        .unwrap_or_default()
-        .to_lowercase()
-}
-
-fn resolved_doc_metadata(
-    doc: &MemoryDocument,
-) -> Option<crate::memory::schema::ResolvedMemoryMetadata> {
-    let workspace_id = doc
-        .metadata
-        .get("namespace")
-        .and_then(|value| value.get("workspace_id"))
-        .and_then(|value| value.as_str())
-        .or_else(|| {
-            doc.metadata
-                .get("workspace_id")
-                .and_then(|value| value.as_str())
-        })
-        .unwrap_or("default");
-    resolve_metadata(&doc.path, &doc.metadata, workspace_id, None).ok()
-}
-
-fn locomo_lexical_score(doc: &MemoryDocument, normalized_query: &str) -> f32 {
-    let content = doc.content.to_lowercase();
-    let path = doc.path.to_lowercase();
-    let terms = locomo_query_terms(normalized_query);
-    let phrases = locomo_phrases(&terms);
-    let speaker = metadata_text_lower(doc, "speaker");
-    let subject = doc
-        .metadata
-        .get("event_subject")
-        .and_then(|value| value.as_str())
-        .unwrap_or_else(|| {
-            doc.metadata
-                .get("speaker")
-                .and_then(|value| value.as_str())
-                .unwrap_or_default()
-        })
-        .to_lowercase();
-    let action = metadata_text_lower(doc, "event_action");
-    let resolved_date = metadata_text_lower(doc, "resolved_date");
-    let normalized_value = metadata_text_lower(doc, "normalized_value");
-    let answer_span = metadata_text_lower(doc, "answer_span");
-    let memory_kind = doc
-        .metadata
-        .get("memory_kind")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default();
-    let resolved = resolved_doc_metadata(doc);
-    let category = doc
-        .metadata
-        .get("category")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default();
-    let temporal_query = is_temporal_query(normalized_query);
-
-    // Generic question pattern detection (context-agnostic)
-    let is_shared_query = normalized_query.contains("in common")
-        || normalized_query.contains("both like")
-        || normalized_query.contains("both have")
-        || normalized_query.contains("what do")
-            && normalized_query.contains("and")
-            && normalized_query.contains("both");
-    let is_why_query = normalized_query.starts_with("why")
-        || normalized_query.contains(" why ")
-        || normalized_query.contains("reason")
-        || normalized_query.contains("because");
-    let is_what_think_query = normalized_query.contains("think")
-        || normalized_query.contains("believe")
-        || normalized_query.contains("opinion")
-        || normalized_query.contains("ideal")
-        || normalized_query.contains("prefer");
-
-    let mut score = 0.0f32;
-    let mut matched_terms = 0usize;
-    for term in &terms {
-        let mut term_score = 0.0f32;
-
-        if !speaker.is_empty() && speaker == **term {
-            term_score += 18.0;
-        }
-        if !subject.is_empty() && subject == **term {
-            term_score += 18.0;
-        }
-        if action.contains(*term) {
-            term_score += 14.0;
-        }
-        if normalized_value.contains(*term) || answer_span.contains(*term) {
-            term_score += 12.0;
-        }
-        if path.contains(*term) {
-            term_score += 8.0;
-        }
-        if content.contains(*term) {
-            term_score += 4.0;
-        }
-
-        if term_score > 0.0 {
-            matched_terms += 1;
-            score += term_score;
-        }
-    }
-
-    score += (matched_terms * matched_terms * 2) as f32;
-
-    for phrase in &phrases {
-        if action.contains(phrase)
-            || normalized_value.contains(phrase)
-            || answer_span.contains(phrase)
-        {
-            score += 18.0;
-        } else if content.contains(phrase) || path.contains(phrase) {
-            score += 9.0;
-        }
-    }
-
-    if normalized_query.split_whitespace().count() >= 2 && content.contains(normalized_query) {
-        score += 10.0;
-    }
-
-    if !speaker.is_empty() && normalized_query.contains(&speaker) {
-        score += 14.0;
-    }
-
-    // Generic context-agnostic scoring patterns
-    // For "shared/common" queries, prioritize documents with matching subjects
-    if is_shared_query {
-        // Boost documents with structured facts (fact_atom, entity_state) for shared queries
-        if matches!(memory_kind, "fact_atom" | "entity_state") {
-            score += 35.0;
-        }
-        // Prioritize normalized_value and answer_span which contain extracted facts
-        if !normalized_value.is_empty() || !answer_span.is_empty() {
-            score += 25.0;
-        }
-        // Penalize summaries for shared queries (prefer primary sources)
-        if memory_kind == "summary_fact" {
-            score *= 0.15;
-        }
-    }
-
-    // For "why" queries, prioritize documents with reason/explanation patterns
-    if is_why_query {
-        // Boost documents that contain causal language
-        let has_reason = content.contains("because")
-            || content.contains("'cause")
-            || content.contains("since")
-            || content.contains("reason")
-            || content.contains("to share")
-            || content.contains("to start")
-            || content.contains("decided")
-            || content.contains("wanted");
-        if has_reason {
-            score += 30.0;
-        }
-        // Boost structured facts with clear values
-        if !normalized_value.is_empty() {
-            score += 20.0;
-        }
-        // Penalize summaries for why queries
-        if memory_kind == "summary_fact" {
-            score *= 0.2;
-        }
-    }
-
-    // For "what think/opinion" queries, prioritize sentiment/opinion content
-    if is_what_think_query {
-        // Boost documents with opinion markers
-        let has_opinion = content.contains("think")
-            || content.contains("believe")
-            || content.contains("feel")
-            || content.contains("prefer")
-            || content.contains("ideal")
-            || content.contains("favorite")
-            || contains_opinion_adjectives(&content);
-        if has_opinion {
-            score += 25.0;
-        }
-        // Boost documents with extracted values
-        if !normalized_value.is_empty() || !answer_span.is_empty() {
-            score += 15.0;
-        }
-        // Penalize summaries for opinion queries
-        if memory_kind == "summary_fact" {
-            score *= 0.2;
-        }
-    }
-
-    if let Some(resolved) = &resolved {
-        match resolved.evidence_kind {
-            Some(EvidenceKind::TemporalEvent) if temporal_query => score += 60.0,
-            Some(
-                EvidenceKind::FactAtom | EvidenceKind::EntityState | EvidenceKind::SummaryFact,
-            ) if !temporal_query => {
-                score += 28.0;
-            }
-            Some(
-                EvidenceKind::FactAtom | EvidenceKind::EntityState | EvidenceKind::SummaryFact,
-            ) => {
-                score += 12.0;
-            }
-            Some(EvidenceKind::SourceTurn) => score += 8.0,
-            _ => {}
-        }
-
-        if let Some(symbol) = resolved.provenance.symbol.as_ref() {
-            if normalized_query.contains(&symbol.to_ascii_lowercase()) {
-                score += 24.0;
-            }
-        }
-        if let Some(file_path) = resolved.provenance.file_path.as_ref() {
-            if normalized_query.contains(&file_path.to_ascii_lowercase()) {
-                score += 16.0;
-            }
-        }
-        if let Some(url) = resolved.provenance.url.as_ref() {
-            if normalized_query.contains(&url.to_ascii_lowercase()) {
-                score += 16.0;
-            }
-        }
-    }
-
-    match memory_kind {
-        "temporal_event" if temporal_query => {
-            score += 60.0;
-        }
-        "fact_atom" | "entity_state" | "summary_fact" if !temporal_query => {
-            score += 28.0;
-        }
-        "fact_atom" | "entity_state" | "summary_fact" => {
-            score += 12.0;
-        }
-        _ => {}
-    }
-
-    if !resolved_date.is_empty() {
-        score += if temporal_query { 24.0 } else { 6.0 };
-        score += match infer_date_granularity(&resolved_date) {
-            "full_date" => 10.0,
-            "month_year" => 6.0,
-            "year" => 2.0,
-            _ => 0.0,
-        };
-    }
-
-    match category {
-        "conversation" => {
-            score += if temporal_query { 18.0 } else { 10.0 };
-        }
-        "observation" => {
-            score += 2.0;
-        }
-        "session_summary" => {
-            score -= if temporal_query { 70.0 } else { 28.0 };
-        }
-        _ => {}
-    }
-
-    if category == "session_summary" && memory_kind.is_empty() {
-        score *= if temporal_query { 0.02 } else { 0.15 };
-    }
-
-    // LOCOMO fix: Boost structured data (pricing, numbers) for factuality queries
-    // Detect pricing/cost/value queries in both English and Spanish
-    let pricing_query = normalized_query.contains("pricing")
-        || normalized_query.contains("price")
-        || normalized_query.contains("precios")
-        || normalized_query.contains("precio")
-        || normalized_query.contains("costo")
-        || normalized_query.contains("coste")
-        || normalized_query.contains("valor")
-        || normalized_query.contains("fee")
-        || normalized_query.contains("tarifa")
-        || normalized_query.contains("cuanto")  // Spanish "how much"
-        || normalized_query.contains("cuál")     // Spanish "which"
-        || normalized_query.contains("cuáles"); // Spanish "which" plural
-
-    if pricing_query {
-        // Boost documents that contain numeric values (likely pricing facts)
-        // Patterns: $499, 499, 499.99, etc.
-        let has_numeric = content.contains('$')
-            || content.contains("/mes")
-            || content.contains("/mo")
-            || content.contains("/month")
-            || content.contains("/monthly")
-            || content.contains("/year")
-            || content.contains("/annual")
-            || regex::Regex::new(r"\d+[.,]?\d*")
-                .map(|re| re.is_match(&content))
-                .unwrap_or(false);
-
-        if has_numeric {
-            score += 30.0;
-        }
-
-        // Extra boost for fact_atom/entity_state with normalized_value
-        // These are extracted structured facts that are most reliable
-        if !normalized_value.is_empty() && has_numeric {
-            score += 25.0;
-        }
-
-        // Boost for tier/version terms (Starter, Pro, Enterprise, etc.)
-        let tier_terms = [
-            "starter",
-            "pro",
-            "enterprise",
-            "basic",
-            "plan",
-            "tier",
-            "version",
-        ];
-        for tier in &tier_terms {
-            if normalized_query.contains(tier) && (content.contains(tier) || path.contains(tier)) {
-                score += 15.0;
-            }
-        }
-    }
-
-    score.max(0.0)
-}
-
-fn lexical_score(doc: &MemoryDocument, normalized_query: &str) -> f32 {
-    if normalized_query.is_empty() {
-        return 0.0;
-    }
-
-    if is_locomo_document(&doc.path, &doc.metadata) {
-        return locomo_lexical_score(doc, normalized_query);
-    }
-
-    let content = doc.content.to_lowercase();
-    let path = doc.path.to_lowercase();
-    let query_terms: Vec<&str> = normalized_query
-        .split_whitespace()
-        .filter(|term| !term.is_empty())
-        .collect();
-    let mut matched_terms = 0usize;
-    let mut score = 0.0f32;
-    for term in &query_terms {
-        let content_hits = content.matches(term).count() as f32;
-        let path_hits = path.matches(term).count() as f32 * 2.0;
-        if content_hits > 0.0 || path_hits > 0.0 {
-            matched_terms += 1;
-        }
-        score += content_hits + path_hits;
-    }
-    score += (matched_terms * matched_terms) as f32;
-
-    let memory_kind = doc
-        .metadata
-        .get("memory_kind")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default();
-    let resolved = resolved_doc_metadata(doc);
-    let category = doc
-        .metadata
-        .get("category")
-        .and_then(|value| value.as_str())
-        .unwrap_or_default();
-
-    if normalized_query.split_whitespace().count() >= 2 && content.contains(normalized_query) {
-        score += 6.0;
-    }
-
-    for (query_signal, content_signal, bonus) in [
-        ("sunrise", "sunrise", 12.0),
-        ("support", "support group", 12.0),
-        ("charity", "charity race", 12.0),
-        ("camping", "camping", 12.0),
-        ("identity", "transgender", 10.0),
-        ("relationship", "single", 10.0),
-        ("research", "adoption agenc", 8.0),
-        ("field", "counsel", 8.0),
-        ("pursue", "counsel", 8.0),
-        ("what", "what", 2.0),
-        ("who", "who", 3.0),
-        ("how", "how", 2.0),
-        ("why", "why", 2.0),
-        ("which", "which", 2.0),
-    ] {
-        if normalized_query.contains(query_signal) && content.contains(content_signal) {
-            score += bonus;
-        }
-    }
-
-    if matches!(
-        memory_kind,
-        "fact_atom" | "entity_state" | "temporal_event" | "summary_fact"
-    ) {
-        score += 5.0;
-    }
-
-    if let Some(resolved) = &resolved {
-        match resolved.kind {
-            MemoryKind::Repo | MemoryKind::File | MemoryKind::Symbol | MemoryKind::Url => {
-                score += 5.0;
-            }
-            MemoryKind::Decision | MemoryKind::Task | MemoryKind::Fact
-                if query_terms.len() >= 2 =>
-            {
-                score += 3.0;
-            }
-            _ => {}
-        }
-
-        if let Some(evidence_kind) = resolved.evidence_kind {
-            match evidence_kind {
-                EvidenceKind::SourceTurn => score += 6.0,
-                EvidenceKind::FactAtom | EvidenceKind::EntityState => score += 8.0,
-                EvidenceKind::TemporalEvent if normalized_query.contains("when") => score += 10.0,
-                EvidenceKind::SessionSummary => score *= 0.5,
-                _ => {}
-            }
-        }
-
-        for exact in [
-            resolved.provenance.symbol.as_ref(),
-            resolved.provenance.file_path.as_ref(),
-            resolved.provenance.repo_url.as_ref(),
-            resolved.provenance.url.as_ref(),
-            resolved.namespace.session_id.as_ref(),
-            resolved.namespace.agent_id.as_ref(),
-            resolved.namespace.user_id.as_ref(),
-            resolved.namespace.project.as_ref(),
-        ]
-        .into_iter()
-        .flatten()
-        {
-            let lowered = exact.to_ascii_lowercase();
-            if !lowered.is_empty() && normalized_query.contains(&lowered) {
-                score += 18.0;
-            }
-        }
-    }
-
-    if doc
-        .metadata
-        .get("normalized_value")
-        .and_then(|value| value.as_str())
-        .is_some()
-    {
-        score += 2.0;
-    }
-
-    match category {
-        "session_summary" => score *= 0.2,
-        "conversation" => score *= 1.2,
-        "observation" => score *= 0.8,
-        _ => {}
-    }
-
-    score
-}
-
-pub(crate) fn cosine_similarity(left: &[f32], right: &[f32]) -> f32 {
-    if left.is_empty() || right.is_empty() || left.len() != right.len() {
-        return 0.0;
-    }
-
-    let dot = left
-        .iter()
-        .zip(right.iter())
-        .map(|(a, b)| a * b)
-        .sum::<f32>();
-    let left_magnitude = left.iter().map(|value| value * value).sum::<f32>().sqrt();
-    let right_magnitude = right.iter().map(|value| value * value).sum::<f32>().sqrt();
-
-    if left_magnitude == 0.0 || right_magnitude == 0.0 {
-        return 0.0;
-    }
-
-    dot / (left_magnitude * right_magnitude)
-}
-
-// ============================================
-// pplx-embed Integration for Semantic Search
-// ============================================
-
-/// Generate embedding using pplx-embed service
-///
-/// CRITICAL FIXES applied:
-/// 1. Embedding cache: Avoids re-embedding identical content (~118ms savings per document)
-/// 2. Exponential backoff: Replaces fixed 250ms retries with 100ms→200ms→400ms→800ms→1600ms (max 2s)
-async fn generate_embedding(text: &str) -> Result<Vec<f32>> {
-    let new_embedder_configured =
-        std::env::var("XAVIER2_EMBEDDER").is_ok() || std::env::var("OPENAI_API_KEY").is_ok();
-    let legacy_embedder_configured = std::env::var("XAVIER2_EMBEDDING_URL").is_ok();
-
-    if !new_embedder_configured && !legacy_embedder_configured {
-        return Ok(Vec::new());
-    }
-    let preprocessed = preprocess_for_embedding(text);
-    let cache_key = embedding_cache_key(&preprocessed);
-
-    // CRITICAL FIX: Check embedding cache first to avoid redundant API calls
-    {
-        let cache = EMBEDDING_CACHE.read().await;
-        if let Some(entry) = cache.get(&cache_key) {
-            // Check if entry is still valid (within TTL)
-            if Instant::now().duration_since(entry.cached_at).as_secs() < EMBEDDING_CACHE_TTL_SECS {
-                tracing::debug!("Embedding cache HIT for key: {}", &cache_key[..16]);
-                return Ok(entry.vector.clone());
-            }
-        }
-    }
-
-    let mut last_error = None;
-    let mut delay_ms: u64 = 100;
-    let max_delay_ms: u64 = 2000;
-
-    if new_embedder_configured {
-        if let Ok(embedder) = crate::embedding::build_embedder_from_env().await {
-            if embedder.dimension() > 0 {
-                for attempt in 0..3 {
-                    match embedder.encode(&preprocessed).await {
-                        Ok(vector) => {
-                            let mut cache = EMBEDDING_CACHE.write().await;
-                            cache.insert(
-                                cache_key,
-                                EmbeddingCacheEntry {
-                                    vector: vector.clone(),
-                                    cached_at: Instant::now(),
-                                },
-                            );
-                            if cache.len() % 10 == 0 {
-                                drop(cache);
-                                clean_embedding_cache().await;
-                            }
-                            return Ok(vector);
-                        }
-                        Err(error) => {
-                            last_error = Some(anyhow::anyhow!(error.to_string()));
-                            if attempt < 2 {
-                                tokio::time::sleep(std::time::Duration::from_millis(delay_ms))
-                                    .await;
-                                delay_ms = (delay_ms * 2).min(max_delay_ms);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    if legacy_embedder_configured {
-        let client = crate::memory::embedder::EmbeddingClient::from_env()?;
-
-        for attempt in 0..3 {
-            match client.embed(&preprocessed).await {
-                Ok(vector) => {
-                    let mut cache = EMBEDDING_CACHE.write().await;
-                    cache.insert(
-                        cache_key,
-                        EmbeddingCacheEntry {
-                            vector: vector.clone(),
-                            cached_at: Instant::now(),
-                        },
-                    );
-                    if cache.len() % 10 == 0 {
-                        drop(cache);
-                        clean_embedding_cache().await;
-                    }
-                    return Ok(vector);
-                }
-                Err(error) => {
-                    last_error = Some(error);
-                    if attempt < 2 {
-                        tokio::time::sleep(std::time::Duration::from_millis(delay_ms)).await;
-                        delay_ms = (delay_ms * 2).min(max_delay_ms);
-                    }
-                }
-            }
-        }
-    }
-
-    Err(last_error.unwrap_or_else(|| anyhow::anyhow!("embedding generation failed")))
-}
-
-pub fn extract_answer(content: &str, category: &str) -> Option<String> {
-    let text = content.trim();
-    if text.is_empty() {
-        return None;
-    }
-
-    match category {
-        "2" => {
-            static DATE_RE: LazyLock<Regex> = LazyLock::new(|| {
-                Regex::new(r"(?i)\b(?:\d{1,2}\s+[A-Za-z]+\s+\d{4}|[A-Za-z]+\s+\d{1,2},\s+\d{4}|(19|20)\d{2})\b").unwrap()
-            });
-            DATE_RE.find(text).map(|m| m.as_str().trim().to_string())
-        }
-        "3" => {
-            let sentence = text
-                .split(['.', '!', '?'])
-                .map(str::trim)
-                .find(|sentence| {
-                    let lowered = sentence.to_lowercase();
-                    [
-                        "think",
-                        "believe",
-                        "feel",
-                        "guess",
-                        "suppose",
-                        "probably",
-                        "definitely",
-                        "maybe",
-                        "opinion",
-                        "view",
-                        "perspective",
-                        "seems",
-                        "appears",
-                        "likely",
-                        "certainly",
-                        "perhaps",
-                        "wonder",
-                    ]
-                    .iter()
-                    .any(|keyword| lowered.contains(keyword))
-                })
-                .or_else(|| {
-                    text.split(['.', '!', '?'])
-                        .map(str::trim)
-                        .find(|s| !s.is_empty())
-                })?;
-            Some(sentence.to_string())
-        }
-        "4" => {
-            let sentence = text
-                .split(['.', '!', '?'])
-                .map(str::trim)
-                .find(|sentence| {
-                    let lowered = sentence.to_lowercase();
-                    [
-                        "decided",
-                        "planning",
-                        "planned",
-                        "will",
-                        "going to",
-                        "intend",
-                        "promised",
-                        "try",
-                        "started",
-                        "beginning",
-                        "began",
-                        "going to start",
-                        "want to",
-                        "hoping to",
-                        "aiming to",
-                    ]
-                    .iter()
-                    .any(|keyword| lowered.contains(keyword))
-                })
-                .or_else(|| {
-                    text.split(['.', '!', '?'])
-                        .map(str::trim)
-                        .find(|s| !s.is_empty())
-                })?;
-            Some(sentence.to_string())
-        }
-        _ => text
-            .split(['.', '!', '?'])
-            .map(str::trim)
-            .find(|sentence| !sentence.is_empty())
-            .map(|sentence| sentence.to_string()),
-    }
-}
-
-/// Speaker-aware text preprocessing for conversation data (LoCoMo benchmark).
-///
-/// Detects speaker patterns like "Caroline:", "[Caroline]", "Speaker: Caroline",
-/// "Person: ..." and prepends a structured speakers summary to help the embedding
-/// model better capture who-said-what, improving retrieval for Q&A over conversations.
-///
-/// ENHANCED: Now includes:
-/// - Speaker turn-taking structure (who speaks when)
-/// - Quoted speech preservation (semantic quotes are often key answers)
-/// - Sequential context (previous speaker matters for multi-turn convos)
-fn preprocess_for_embedding(text: &str) -> String {
-    let speakers = extract_speakers(text);
-
-    if speakers.is_empty() {
-        // Still preprocess to handle quoted speech
-        return preserve_quoted_speech(text);
-    }
-
-    // Build structured speaker context with turn-taking info
-    let speaker_list: Vec<String> = speakers.iter().map(|s| format!("[{}]", s)).collect();
-
-    // Preserve quoted speech which often contains answers
-    let text_with_quotes = preserve_quoted_speech(text);
-
-    let speaker_ctx = format!(
-        "Conversation between: {}. \nQuote context: {}\n\n",
-        speaker_list.join(", "),
-        speaker_list.join(" said, ")
-    );
-    format!("{}{}", speaker_ctx, text_with_quotes)
-}
-
-/// Preserve quoted speech as special markers since quotes often contain key answers
-fn preserve_quoted_speech(text: &str) -> String {
-    // Replace quoted text with a marker to emphasize it in embeddings
-    let mut result = text.to_string();
-
-    // Pattern for quoted speech: "..." or '...'
-    let quote_re = regex::Regex::new(r#"["']([^"']+)["']"#).unwrap();
-
-    let mut quote_count = 0;
-    result = quote_re
-        .replace_all(&result, |caps: &regex::Captures| {
-            quote_count += 1;
-            let quote = &caps[1];
-            format!("[QUOTE{}: {}]", quote_count, quote)
-        })
-        .to_string();
-
-    result
-}
-
-/// Extract unique speaker names from conversation text.
-fn extract_speakers(text: &str) -> Vec<String> {
-    let mut speakers = Vec::new();
-    let mut seen = std::collections::HashSet::new();
-
-    for re in &[&*SPEAKER_COLON_RE, &*SPEAKER_BRACKET_RE, &*SPEAKER_ROLE_RE] {
-        for cap in re.captures_iter(text) {
-            if let Some(name) = cap.get(1) {
-                let name = name.as_str().trim();
-                if is_likely_speaker(name) && seen.insert(name.to_lowercase()) {
-                    speakers.push(name.to_string());
-                }
-            }
-        }
-    }
-    speakers
-}
-
-/// Extract potential speaker from a question query.
-fn extract_speaker_from_query(query: &str) -> Option<String> {
-    QUERY_SPEAKER_RE.captures(query).and_then(|cap| {
-        let name = cap.get(1)?.as_str();
-        if is_likely_speaker(name) {
-            Some(name.to_string())
-        } else {
-            None
-        }
-    })
-}
-
-/// Heuristic for gendered names to resolve pronouns.
-fn is_female_name(name: &str) -> bool {
-    let name = name.to_lowercase();
-    let female_names = [
-        "caroline",
-        "alice",
-        "sarah",
-        "emma",
-        "olivia",
-        "sophia",
-        "isabella",
-        "mia",
-        "charlotte",
-        "amelia",
-        "mary",
-        "patricia",
-        "jennifer",
-        "linda",
-        "elizabeth",
-        "barbara",
-        "susan",
-        "jessica",
-        "karen",
-    ];
-    female_names.contains(&name.as_str())
-}
-
-fn is_male_name(name: &str) -> bool {
-    let name = name.to_lowercase();
-    let male_names = [
-        "james",
-        "robert",
-        "john",
-        "michael",
-        "david",
-        "william",
-        "richard",
-        "joseph",
-        "thomas",
-        "christopher",
-        "charles",
-        "daniel",
-        "matthew",
-        "anthony",
-        "mark",
-        "donald",
-        "steven",
-        "paul",
-        "andrew",
-        "joshua",
-    ];
-    male_names.contains(&name.as_str())
-}
-
-/// Resolve pronouns "he/she" to names if only one candidate is present.
-fn resolve_pronouns(query: &str, speakers: &[String]) -> String {
-    let mut resolved = query.to_string();
-
-    // Resolve "she"
-    if query.to_lowercase().contains("she") {
-        let female_candidates: Vec<_> = speakers.iter().filter(|s| is_female_name(s)).collect();
-        if female_candidates.len() == 1 {
-            resolved = SHE_RE
-                .replace_all(&resolved, female_candidates[0])
-                .to_string();
-        }
-    }
-
-    // Resolve "he"
-    if query.to_lowercase().contains("he") {
-        let male_candidates: Vec<_> = speakers.iter().filter(|s| is_male_name(s)).collect();
-        if male_candidates.len() == 1 {
-            resolved = HE_RE.replace_all(&resolved, male_candidates[0]).to_string();
-        }
-    }
-
-    resolved
-}
-
-/// Heuristic: a token is likely a speaker name if it's:
-/// - 2-20 characters
-/// - Starts with uppercase
-/// - Contains only letters (possibly with hyphens for hyphenated names)
-/// - Not a common word
-fn is_likely_speaker(s: &str) -> bool {
-    let s = s.trim();
-    if s.len() < 2 || s.len() > 20 {
-        return false;
-    }
-    if !s
-        .chars()
-        .next()
-        .map(|c| c.is_ascii_uppercase())
-        .unwrap_or(false)
-    {
-        return false;
-    }
-    if !s
-        .chars()
-        .all(|c| c.is_ascii_alphabetic() || c == '-' || c == '\'')
-    {
-        return false;
-    }
-    // Filter out common English words and pronouns
-    let common: &[&str] = &[
-        "The", "This", "That", "They", "Then", "There", "Here", "Hello", "Thanks", "Please",
-        "Sorry", "Who", "What", "When", "Where", "Why", "How", "Did", "Was", "Were", "Are", "She",
-        "He", "It",
-    ];
-    if common.contains(&s) {
-        return false;
-    }
-    true
-}
-
-/// Query with semantic search using embeddings
-pub async fn query_with_embedding(
-    memory: &QmdMemory,
-    query_text: &str,
-    limit: usize,
-) -> Result<Vec<MemoryDocument>> {
-    query_with_embedding_filtered(memory, query_text, limit, None).await
-}
-
-pub async fn query_with_embedding_filtered(
-    memory: &QmdMemory,
-    query_text: &str,
-    limit: usize,
-    filters: Option<&MemoryQueryFilters>,
-) -> Result<Vec<MemoryDocument>> {
-    let mut processed_query = query_text.to_string();
-
-    // 1. Extract all speakers currently in memory to assist with pronoun resolution
-    let all_docs = memory.all_documents().await;
-    let mut all_speakers = std::collections::HashSet::new();
-    let locomo_only = !all_docs.is_empty()
-        && all_docs
-            .iter()
-            .all(|doc| is_locomo_document(&doc.path, &doc.metadata));
-    for doc in &all_docs {
-        for speaker in extract_speakers(&doc.content) {
-            all_speakers.insert(speaker);
-        }
-    }
-    let speakers_list: Vec<String> = all_speakers.into_iter().collect();
-
-    // 2. Resolve pronouns if applicable
-    if !speakers_list.is_empty() {
-        processed_query = resolve_pronouns(&processed_query, &speakers_list);
-    }
-
-    // 3. If a name is explicitly mentioned in the query after an interrogative,
-    // ensure it's prioritized in the final retrieval by prepending it to the query.
-    // This dramatically improves semantic matching for "Who did X?" style questions.
-    if let Some(target_speaker) = extract_speaker_from_query(query_text) {
-        // Prepend speaker name for better semantic focus
-        if !processed_query.contains(&target_speaker) {
-            processed_query = format!("{} {}", target_speaker, processed_query);
-        }
-    }
-
-    if locomo_only {
-        return memory
-            .query_filtered(&processed_query, Vec::new(), limit, filters)
-            .await;
-    }
-
-    let query_vector = generate_embedding(&processed_query).await?;
-
-    if query_vector.is_empty() {
-        // Fallback to keyword search with the processed query
-        return memory
-            .search_with_cache_filtered(&processed_query, limit, filters)
-            .await
-            .map(|r| r.documents);
-    }
-
-    // 4. ENHANCED: Use top semantic results to expand query context
-    // Get initial semantic results to understand what the query is about
-    let initial_results = memory
-        .vsearch(query_vector.clone(), 3)
-        .await
-        .unwrap_or_default();
-
-    // If we found relevant documents, create an expanded query that includes
-    // context from those documents to improve recall
-    if !initial_results.is_empty() {
-        let mut context_terms = Vec::new();
-
-        // Extract meaningful terms from top results (avoiding common words)
-        let common_words: std::collections::HashSet<&str> = std::collections::HashSet::from_iter([
-            "the", "a", "an", "is", "are", "was", "were", "be", "been", "being", "have", "has",
-            "had", "do", "does", "did", "will", "would", "could", "should", "may", "might", "must",
-            "shall", "can", "need", "dare", "to", "of", "in", "for", "on", "with", "at", "by",
-            "from", "as", "into", "through", "during", "before", "after", "above", "below", "that",
-            "this", "these", "those", "it", "its", "they", "them", "what", "which", "who", "whom",
-            "whose", "where", "when", "why", "how",
-        ]);
-
-        for doc in initial_results.iter().take(2) {
-            for word in doc.content.split_whitespace() {
-                let w_clean = word
-                    .trim_matches(|c: char| !c.is_alphanumeric())
-                    .to_lowercase();
-                if w_clean.len() >= 4
-                    && !common_words.contains(w_clean.as_str())
-                    && !processed_query.to_lowercase().contains(&w_clean)
-                {
-                    context_terms.push(w_clean);
-                }
-            }
-        }
-
-        // Add context terms to query if we have few results
-        if context_terms.len() >= 2 {
-            let expanded_query = format!("{} {}", processed_query, context_terms.join(" "));
-            // Generate a second embedding with expanded context
-            if let Ok(expanded_vector) = generate_embedding(&expanded_query).await {
-                if !expanded_vector.is_empty() {
-                    // Use the expanded vector for better semantic matching
-                    return memory
-                        .query_filtered(&expanded_query, expanded_vector, limit, filters)
-                        .await;
-                }
-            }
-        }
-    }
-
-    memory
-        .query_filtered(&processed_query, query_vector, limit, filters)
-        .await
-}
-
-/// Deduplicate search results by content hash, keeping most recent by updated_at.
-/// Groups documents by SHA256 content hash, then selects the document with the latest
-/// `updated_at` metadata field (or `created_at` as fallback). Preserves original order
-/// by keeping the first occurrence of each hash group.
-#[allow(dead_code)]
-fn _deduplicate_by_content_hash(results: Vec<MemoryDocument>) -> Vec<MemoryDocument> {
-    use std::collections::HashMap;
-
-    // Group by content hash, tracking (document, latest_updated_at, original_index)
-    let mut hash_groups: HashMap<String, (MemoryDocument, Option<String>, usize)> = HashMap::new();
-    for (idx, doc) in results.into_iter().enumerate() {
-        let content_hash = _compute_content_hash(&doc.content);
-        let updated_at = doc
-            .metadata
-            .get("updated_at")
-            .and_then(|v| v.as_str())
-            .or_else(|| doc.metadata.get("created_at").and_then(|v| v.as_str()))
-            .map(str::to_string);
-
-        hash_groups
-            .entry(content_hash)
-            .and_modify(|(existing_doc, existing_updated, existing_idx)| {
-                // Keep the one with more recent updated_at, or keep existing if tie
-                let is_newer = match (updated_at.as_ref(), existing_updated.as_ref()) {
-                    (Some(new), Some(old)) => new > old,
-                    (Some(_), None) => true, // new has timestamp, existing doesn't
-                    (None, Some(_)) => false, // existing has timestamp, new doesn't
-                    (None, None) => idx < *existing_idx, // tie-break by original order
-                };
-                if is_newer {
-                    *existing_doc = doc.clone();
-                    *existing_updated = updated_at.clone();
-                    *existing_idx = idx;
-                }
-            })
-            .or_insert((doc, updated_at.clone(), idx));
-    }
-
-    // Extract deduplicated results, sort by original order (first occurrence per hash)
-    let mut deduped: Vec<(usize, MemoryDocument)> = hash_groups
-        .into_values()
-        .map(|(doc, _, idx)| (idx, doc))
-        .collect();
-    deduped.sort_by(|a, b| a.0.cmp(&b.0));
-    deduped.into_iter().map(|(_, doc)| doc).collect()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::env;
-
-    #[tokio::test]
-    async fn repeated_searches_hit_cache() {
-        let memory = QmdMemory::new(Arc::new(AsyncRwLock::new(Vec::new())));
-        memory
-            .add_document(
-                "docs/cache".to_string(),
-                "cache acceleration for repeated searches".to_string(),
-                serde_json::json!({}),
-            )
-            .await
-            .unwrap();
-
-        let first = memory
-            .search_with_cache("cache acceleration", 5)
-            .await
-            .unwrap();
-        let second = memory
-            .search_with_cache("cache acceleration", 5)
-            .await
-            .unwrap();
-        let metrics = memory.cache_metrics().await;
-
-        assert!(!first.cache_hit);
-        assert!(second.cache_hit);
-        assert_eq!(metrics.misses, 1);
-        assert_eq!(metrics.hits, 1);
-        assert_eq!(metrics.entries, 1);
-    }
-
-    #[tokio::test]
-    async fn mutating_memory_invalidates_cache() {
-        let memory = QmdMemory::new(Arc::new(AsyncRwLock::new(Vec::new())));
-        memory
-            .add_document(
-                "docs/original".to_string(),
-                "performance tuning for xavier2".to_string(),
-                serde_json::json!({}),
-            )
-            .await
-            .unwrap();
-
-        let _ = memory.search_with_cache("performance", 5).await.unwrap();
-        assert_eq!(memory.cache_metrics().await.entries, 1);
-
-        memory
-            .add_document(
-                "docs/new".to_string(),
-                "new performance tuning guide".to_string(),
-                serde_json::json!({}),
-            )
-            .await
-            .unwrap();
-
-        assert_eq!(memory.cache_metrics().await.entries, 0);
-    }
-
-    #[tokio::test]
-    async fn add_document_skips_embedding_when_service_not_configured() {
-        unsafe {
-            env::remove_var("XAVIER2_EMBEDDING_URL");
-        }
-
-        let memory = QmdMemory::new(Arc::new(AsyncRwLock::new(Vec::new())));
-        memory
-            .add_document(
-                "docs/offline".to_string(),
-                "offline startup should not require embeddings".to_string(),
-                serde_json::json!({ "source": "test" }),
-            )
-            .await
-            .unwrap();
-
-        let stored = memory.get("docs/offline").await.unwrap().unwrap();
-        assert!(stored.embedding.is_empty());
-    }
-
-    #[tokio::test]
-    async fn add_document_creates_clean_locomo_derivatives() {
-        let memory = QmdMemory::new(Arc::new(AsyncRwLock::new(Vec::new())));
-        memory
-            .add_document(
-                "locomo/conv-26/session_1/D1:17".to_string(),
-                "Caroline: I've been researching adoption agencies lately.".to_string(),
-                serde_json::json!({
-                    "benchmark": "locomo",
-                    "speaker": "Caroline",
-                    "session_time": "8 May, 2023"
-                }),
-            )
-            .await
-            .unwrap();
-
-        let stored = memory.all_documents().await;
-        assert!(stored.len() > 1);
-        let derived = stored
-            .iter()
-            .find(|doc| {
-                doc.metadata.get("memory_kind").and_then(|v| v.as_str()) == Some("fact_atom")
-            })
-            .expect("derived fact atom");
-        assert_eq!(
-            derived
-                .metadata
-                .get("normalized_value")
-                .and_then(|v| v.as_str()),
-            Some("Adoption agencies")
-        );
-        assert!(!derived.content.contains("source_path"));
-    }
-
-    #[tokio::test]
-    async fn locomo_search_prioritizes_temporal_derivatives_over_session_summaries() {
-        let memory = QmdMemory::new(Arc::new(AsyncRwLock::new(Vec::new())));
-        memory
-            .add_document(
-                "locomo/conv-26/session_1/summary".to_string(),
-                "Caroline and Melanie spoke on 8 May, 2023. Caroline discussed several LGBTQ experiences and many other summer memories.".to_string(),
-                serde_json::json!({
-                    "benchmark": "locomo",
-                    "session_time": "1:56 pm on 8 May, 2023",
-                    "category": "session_summary",
-                }),
-            )
-            .await
-            .unwrap();
-        memory
-            .add_document(
-                "locomo/conv-26/session_1/D1:3".to_string(),
-                "Caroline: I went to a LGBTQ support group yesterday and it was so powerful."
-                    .to_string(),
-                serde_json::json!({
-                    "benchmark": "locomo",
-                    "session_time": "1:56 pm on 8 May, 2023",
-                    "speaker": "Caroline",
-                    "category": "conversation",
-                }),
-            )
-            .await
-            .unwrap();
-
-        let results = memory
-            .search("When did Caroline go to the LGBTQ support group?", 5)
-            .await
-            .unwrap();
-
-        assert!(!results.is_empty());
-        assert_eq!(
-            results[0]
-                .metadata
-                .get("memory_kind")
-                .and_then(|value| value.as_str()),
-            Some("temporal_event")
-        );
-        assert_eq!(
-            results[0]
-                .metadata
-                .get("resolved_date")
-                .and_then(|value| value.as_str()),
-            Some("7 May 2023")
-        );
-    }
-
-    #[tokio::test]
-    async fn add_document_normalizes_locomo_dia_ids_for_primary_and_derived_docs() {
-        let memory = QmdMemory::new(Arc::new(AsyncRwLock::new(Vec::new())));
-        memory
-            .add_document(
-                "locomo/conv-26/session_1/D1:03".to_string(),
-                "Caroline: I went to a LGBTQ support group yesterday and it was so powerful."
-                    .to_string(),
-                serde_json::json!({
-                    "benchmark": "locomo",
-                    "speaker": "Caroline",
-                    "session_time": "1:56 pm on 8 May, 2023",
-                    "dia_id": "d1:03",
-                    "category": "conversation",
-                }),
-            )
-            .await
-            .unwrap();
-
-        let stored = memory.all_documents().await;
-        let primary = stored
-            .iter()
-            .find(|doc| doc.path == "locomo/conv-26/session_1/D1:03")
-            .expect("primary locomo document");
-        assert_eq!(
-            primary
-                .metadata
-                .get("normalized_dia_id")
-                .and_then(|value| value.as_str()),
-            Some("D1:3")
-        );
-        assert_eq!(
-            primary
-                .metadata
-                .get("dia_id")
-                .and_then(|value| value.as_str()),
-            Some("D1:3")
-        );
-
-        let derived = stored
-            .iter()
-            .find(|doc| doc.path.ends_with("#derived/temporal_event/0"))
-            .expect("derived temporal event");
-        assert_eq!(
-            derived
-                .metadata
-                .get("source_path")
-                .and_then(|value| value.as_str()),
-            Some("locomo/conv-26/session_1/D1:3")
-        );
-        assert_eq!(
-            derived
-                .metadata
-                .get("source_dia_id")
-                .and_then(|value| value.as_str()),
-            Some("D1:3")
-        );
-    }
-
-    #[tokio::test]
-    async fn hybrid_search_uses_rrf_to_combine_keyword_and_vector_hits() {
-        let memory = QmdMemory::new(Arc::new(AsyncRwLock::new(Vec::new())));
-        memory
-            .add(MemoryDocument {
-                id: Some("kw-doc".to_string()),
-                path: "docs/keyword".to_string(),
-                content: "Alice moved to Paris in 2020 to work as a software engineer.".to_string(),
-                metadata: serde_json::json!({}),
-                content_vector: Some(vec![0.0, 1.0]),
-                embedding: vec![0.0, 1.0],
-            })
-            .await
-            .unwrap();
-        memory
-            .add(MemoryDocument {
-                id: Some("semantic-doc".to_string()),
-                path: "docs/semantic".to_string(),
-                content:
-                    "Alice's favorite programming language is Rust, which she learned in 2021."
-                        .to_string(),
-                metadata: serde_json::json!({}),
-                content_vector: Some(vec![1.0, 0.0]),
-                embedding: vec![1.0, 0.0],
-            })
-            .await
-            .unwrap();
-        memory
-            .add(MemoryDocument {
-                id: Some("noise-doc".to_string()),
-                path: "docs/noise".to_string(),
-                content: "Bob studied design and architecture in Boston.".to_string(),
-                metadata: serde_json::json!({}),
-                content_vector: Some(vec![0.0, 0.2]),
-                embedding: vec![0.0, 0.2],
-            })
-            .await
-            .unwrap();
-
-        let results = memory
-            .query_with_hybrid_search("Where did Alice move in 2020?", vec![1.0, 0.0], 3)
-            .await
-            .unwrap();
-
-        let paths: Vec<&str> = results.iter().map(|doc| doc.path.as_str()).collect();
-        assert!(paths.iter().take(2).any(|path| *path == "docs/keyword"));
-        assert!(paths.iter().take(2).any(|path| *path == "docs/semantic"));
-    }
-
-    #[test]
-    fn test_extract_speakers() {
-        let text = "Caroline: Hello\n[James]: Hi\nSpeaker: Alice\nPerson: Robert\nGuest: Emma";
-        let speakers = extract_speakers(text);
-        assert!(speakers.contains(&"Caroline".to_string()));
-        assert!(speakers.contains(&"James".to_string()));
-        assert!(speakers.contains(&"Alice".to_string()));
-        assert!(speakers.contains(&"Robert".to_string()));
-        assert!(speakers.contains(&"Emma".to_string()));
-    }
-
-    #[test]
-    fn test_extract_speaker_from_query() {
-        assert_eq!(
-            extract_speaker_from_query("Who is Caroline?"),
-            Some("Caroline".to_string())
-        );
-        assert_eq!(
-            extract_speaker_from_query("What did James say?"),
-            Some("James".to_string())
-        );
-        assert_eq!(
-            extract_speaker_from_query("When was Alice there?"),
-            Some("Alice".to_string())
-        );
-        assert_eq!(
-            extract_speaker_from_query("Where is Robert?"),
-            Some("Robert".to_string())
-        );
-        assert_eq!(
-            extract_speaker_from_query("Why did Emma laugh?"),
-            Some("Emma".to_string())
-        );
-    }
-
-    #[test]
-    fn test_resolve_pronouns() {
-        let speakers = vec!["Caroline".to_string(), "James".to_string()];
-
-        // Single female candidate
-        assert_eq!(
-            resolve_pronouns("What did she say?", &speakers),
-            "What did Caroline say?"
-        );
-
-        // Single male candidate
-        assert_eq!(
-            resolve_pronouns("What did he say?", &speakers),
-            "What did James say?"
-        );
-
-        // Multiple female candidates - no resolution
-        let speakers_multiple = vec!["Caroline".to_string(), "Alice".to_string()];
-        assert_eq!(
-            resolve_pronouns("What did she say?", &speakers_multiple),
-            "What did she say?"
-        );
-    }
-
-    #[test]
-    fn test_is_likely_speaker() {
-        assert!(is_likely_speaker("Caroline"));
-        assert!(is_likely_speaker("James"));
-        assert!(!is_likely_speaker("Who"));
-        assert!(!is_likely_speaker("What"));
-        assert!(!is_likely_speaker("She"));
-        assert!(!is_likely_speaker("The"));
-    }
-}
